wip(iter-247): split output.rs (3744 lines) into output/ submodules
