+++ crates/hyalo-cli/src/output.rs
@@ -1,8 +1,7 @@
 use std::collections::HashMap;
 use std::fmt::Write as _;
 
-use jaq_core::load::{self, Arena, File, Loader};
-use jaq_core::{Compiler, Ctx, Native, Vars, data};
+use jaq_core::{Native, data};
 use jaq_json::Val;
 use serde::Serialize;
 use serde_json::json;
@@ -430,3323 +429,36 @@ pub fn format_budget_error(
 }
 
 // ---------------------------------------------------------------------------
-// jq filter constants — one per output type
+// Submodules (iteration 247)
 // ---------------------------------------------------------------------------
-
-/// `PropertyInfo` (used by `--fields properties-typed`): `{name, type, value}`
-/// When value is an array (list type), join elements with ", " for readability.
-const PROPERTY_INFO_FILTER: &str = r#""\(.name) (\(.type)): \(if (.value | type) == "array" then "[" + (.value | join(", ")) + "]" else .value end)""#;
-
-/// `PropertySummaryEntry`: `{count, name, type, mixed_types?}`
-///
-/// When `mixed_types` is present the type column shows `mixed (N text, M number…)`
-/// so the user can see at a glance that the property is not uniformly typed.
-const PROPERTY_SUMMARY_ENTRY_FILTER: &str = r#""\(.name)\t\(if .mixed_types then "mixed (" + (.mixed_types | map("\(.count) \(.type)") | join(", ")) + ")" else .type end)\t\(.count) \(if .count == 1 then "file" else "files" end)""#;
-
-/// `TagSummary`: `{tags, total}`
-const TAG_SUMMARY_FILTER: &str = r#""\(.total) unique \(if .total == 1 then "tag" else "tags" end)\n\(.tags | map("  \(.name)\t\(.count) \(if .count == 1 then "file" else "files" end)") | join("\n"))""#;
-
-/// `TagSummaryEntry`: `{count, name}`
-const TAG_SUMMARY_ENTRY_FILTER: &str =
-    r#""\(.name)\t\(.count) \(if .count == 1 then "file" else "files" end)""#;
-
-// Every `LinkInfo` text rendering opens with its source line (iter-215,
-// dogfood UX-6): `  line 12: …`. `.line` is always present on a `LinkInfo`,
-// but JSON written by a pre-215 hyalo (or a hand-built fixture) may omit it,
-// so the `// 0` fallbacks in each filter keep those renderable instead of
-// failing. `line 0` is the "unknown" rendering — no real source line is 0,
-// they are 1-based.
-
-/// `LinkInfo` — just target: `{line, target}`
-/// Format: `  line 12: "target" (unresolved)`
-const LINK_INFO_TARGET_FILTER: &str = r#""  line \(.line // 0): \"\(.target)\" (unresolved)""#;
-
-/// `LinkInfo` with path: `{line, path, target}`
-/// Format: `  line 12: "target" → "path"`
-const LINK_INFO_PATH_FILTER: &str = r#""  line \(.line // 0): \"\(.target)\" → \"\(.path)\"""#;
-
-/// `LinkInfo` with label: `{label, line, target}`
-/// Format: `  line 12: "target" (unresolved) [label]`
-const LINK_INFO_LABEL_FILTER: &str =
-    r#""  line \(.line // 0): \"\(.target)\" (unresolved) [\(.label)]""#;
-
-/// `LinkInfo` with path and label: `{label, line, path, target}`
-/// Format: `  line 12: "target" → "path" [label]`
-const LINK_INFO_FULL_FILTER: &str =
-    r#""  line \(.line // 0): \"\(.target)\" → \"\(.path)\" [\(.label)]""#;
-
-/// `LinkInfo` carrying a `#fragment` (L-21, iter-190).
-///
-/// Unified filter for every anchored-link shape (with/without `path`, `label`,
-/// and `broken_anchor`). Missing optional keys read as `null` in jq, so a
-/// single filter covers all six fragment-bearing key signatures:
-/// - `"target"` and `"path"` render as before, with the `#fragment` appended;
-/// - `broken_anchor == true` marks the anchor as `(broken anchor)` after the
-///   (resolved) path, so it reads distinctly from a broken TARGET
-///   `(unresolved)`;
-/// - a `[label]` suffix is appended when present.
-const LINK_INFO_ANCHORED_FILTER: &str = r#""  line \(.line // 0): \"\(.target)#\(.fragment)\"\(if .path then " → \"\(.path)\"\(if .broken_anchor then " (broken anchor)" else "" end)" else " (unresolved)" end)\(if .label then " [\(.label)]" else "" end)""#;
-
-/// `LinkInfo` whose target resolves above the vault root (iter-193).
-///
-/// One filter for every out-of-vault shape: `path` is always absent, while
-/// `fragment` and `label` are optional and read as `null` when missing.
-/// Renders `(out of vault)` rather than `(unresolved)` so a target that is
-/// merely out of scope never reads as a broken link.
-const LINK_INFO_OUT_OF_VAULT_FILTER: &str = r##""  line \(.line // 0): \"\(.target)\(if .fragment then "#\(.fragment)" else "" end)\" (out of vault)\(if .label then " [\(.label)]" else "" end)""##;
-
-/// `TaskCount`: `{done, total}`
-const TASK_COUNT_FILTER: &str = r#""[\(.done)/\(.total)]""#;
-
-/// `OutlineSection` without tasks: `{code_blocks, heading, level, line, links}`
-const OUTLINE_SECTION_FILTER: &str = r##""\("#" * .level) \(.heading // "(pre-heading)")\(if (.links | length) > 0 then "\n\(.links | map("  → \"\(.)\"") | join("\n"))" else "" end)""##;
-
-/// `OutlineSection` with tasks: `{code_blocks, heading, level, line, links, tasks}`
-///
-/// NEW-16 (dogfood pre3): `.heading` is the raw markdown text, which may
-/// itself already carry a hand-written `[n/m]` count (`## Tasks [6/6]`) —
-/// unconditionally appending the computed one doubled it (`## Tasks [6/6]
-/// [1/2]`), with the hand-written half free to go stale the moment a task
-/// checkbox changed. Strips any such trailing bracket count from the heading
-/// before appending the one hyalo actually computed, so it renders once and
-/// is always current.
-///
-/// PR #251 review N15: this filter only ever runs on a section that DOES
-/// have `.tasks` (see the `code_blocks,heading,level,line,links,tasks`
-/// dispatch key below), so the strip always fires — there is no "no tasks,
-/// keep the text" branch here the way `build_file_object_filter`'s inline
-/// sections filter has. A heading that ends in bracket text shaped like
-/// `[n/m]` for a reason *other* than a task count (`## Aspect Ratio [16/9]`,
-/// and it happens to have a task list under it) loses that text: the strip
-/// is a plain regex match on shape, not semantics, and cannot tell "this is
-/// a stale task count" from "this looks like one." Accepted trade-off — the
-/// doubled/stale-count case this filter exists to fix is far more common
-/// than a coincidental `[n/m]`-shaped heading suffix on a section that also
-/// contains a task list.
-const OUTLINE_SECTION_WITH_TASKS_FILTER: &str = r##""\("#" * .level) \((.heading // "(pre-heading)") | sub("\\s*\\[[0-9]+/[0-9]+\\]\\s*$"; "")) [\(.tasks.done)/\(.tasks.total)]\(if (.links | length) > 0 then "\n\(.links | map("  → \"\(.)\"") | join("\n"))" else "" end)""##;
-
-/// `TaskInfo`: `{done, line, status, text}`
-const TASK_INFO_FILTER: &str =
-    r#""line \(.line): [\(.status)] \(.text)\(if .done then " (done)" else "" end)""#;
-
-/// `TaskReadResult`: `{done, file, line, status, text}`
-const TASK_READ_RESULT_FILTER: &str =
-    r#""\"\(.file)\":\(.line) [\(.status)] \(.text)\(if .done then " (done)" else "" end)""#;
-
-/// `TaskDryRunResult`: `{done, file, line, old_status, status, text}`
-/// Format: `"file":line [old] -> [new] text` — makes the direction of change
-/// explicit for `task toggle --dry-run`.
-const TASK_DRY_RUN_RESULT_FILTER: &str =
-    r#""\"\(.file)\":\(.line) [\(.old_status)] -> [\(.status)] \(.text)""#;
-
-/// `VaultSummary`: `{dir, dead_ends, files, links, orphans, properties, recent_files, status, tags, tasks}`
-/// Compact single-line-per-section format (~20-30 lines regardless of vault size).
-///
-/// iter-247: `.dir` is deliberately **not** rendered here. It used to lead the
-/// text report as a `kb dir: <path>` banner — the only command that prefixed its
-/// text output with the vault it had just resolved — which put a
-/// cwd-dependent path on stdout for anyone scripting `--format text`. The vault
-/// dir now goes to stderr as a `note:` (see `commands::summary::summary`),
-/// matching how `--dir`-switching and other resolution context is already
-/// announced, and stays in the JSON payload as `.dir` for machine consumers.
-const VAULT_SUMMARY_FILTER: &str = r#""Files: \(.files.total)\nDirectories: \(if (.files.directories | length) > 0 then (.files.directories | .[:7] | map("\(.directory)/ (\(.count))") | join(", ")) + (if (.files.directories | length) > 7 then ", ..." else "" end) else "(none)" end)\nProperties: \(.properties | length) — \(if (.properties | length) > 0 then (.properties | sort_by(-.count) | .[:7] | map("\(.name) (\(.count))") | join(", ")) + (if (.properties | length) > 7 then ", ..." else "" end) else "(none)" end)\nTags: \(.tags.total) — \(if (.tags.tags | length) > 0 then (.tags.tags | .[:7] | map("\(.name) (\(.count))") | join(", ")) + (if (.tags.tags | length) > 7 then ", ..." else "" end) else "(none)" end)\nTasks: \(.tasks.done)/\(.tasks.total)\nLinks: \(.links.total) total, \(.links.broken) broken\(if .links.broken_anchors > 0 then ", \(.links.broken_anchors) broken anchor\(if .links.broken_anchors == 1 then "" else "s" end)" else "" end)\(if .links.out_of_vault > 0 then ", \(.links.out_of_vault) out of vault" else "" end)\nOrphans: \(.orphans)\nDead-ends: \(.dead_ends)\nStatus: \(if (.status | length) > 0 then (.status | sort_by(-.count) | map("\(.value) (\(.count))") | join(", ")) else "(none)" end)\nRecent: \(if (.recent_files | length) > 0 then (.recent_files | map(.path) | join(", ")) else "(none)" end)""#;
-
-/// `FindTaskInfo`: `{done, line, section, status, text}`
-/// Format: `  [x] text (line N, section)` or `  [ ] text (line N, section)`
-const FIND_TASK_INFO_FILTER: &str =
-    r#""  [\(if .done then "x" else " " end)] \(.text) (line \(.line), \(.section))""#;
-
-/// `ContentMatch`: `{line, section, text}`
-/// Format: `  line N (section): text`
-const CONTENT_MATCH_FILTER: &str = r#""  line \(.line) (\(.section)): \(.text)""#;
-
-/// Mutation result with `property` + `value` fields:
-/// covers `SetPropertyResult`, `AppendPropertyResult`, and `RemovePropertyResult` (with value).
-/// Key signature: `dry_run,modified,property,scanned,skipped,total,value` (without note)
-/// or `dry_run,modified,note,property,scanned,skipped,total,value` (with note, alphabetically sorted).
-/// Format: `[dry-run] property=value: N/T modified (S scanned)` when dry-run; omits prefix otherwise.
-/// Appends `(S scanned)` when not all scanned files were processed (e.g. where-filters).
-/// Appends `  note: <msg>` when a `note` field is present.
-const PROPERTY_VALUE_MUTATION_FILTER: &str = r#""\(if .dry_run then "[dry-run] " else "" end)\(.property)=\(.value): \(.modified | length)/\(.total) modified\(if .scanned != .total then " (\(.scanned) scanned)" else "" end)\(if (.modified | length) > 0 then "\n\(.modified | map("  \"\(.)\"") | join("\n"))" else "" end)\(if .note then "\n  note: \(.note)" else "" end)""#;
-
-/// Mutation result with `property` only (no value field):
-/// covers `RemovePropertyResult` (without value).
-/// Key signature: `dry_run,modified,property,scanned,skipped,total`
-/// Format: `[dry-run] property: N/T modified (S scanned)` when dry-run; omits prefix otherwise.
-/// Appends `(S scanned)` when not all scanned files were processed (e.g. where-filters).
-const PROPERTY_MUTATION_FILTER: &str = r#""\(if .dry_run then "[dry-run] " else "" end)\(.property): \(.modified | length)/\(.total) modified\(if .scanned != .total then " (\(.scanned) scanned)" else "" end)\(if (.modified | length) > 0 then "\n\(.modified | map("  \"\(.)\"") | join("\n"))" else "" end)""#;
-
-/// Mutation result with `tag` field:
-/// covers `SetTagResult` and `RemoveTagResult`.
-/// Key signature: `dry_run,modified,scanned,skipped,tag,total`
-/// Format: `[dry-run] tag: N/T modified (S scanned)` when dry-run; omits prefix otherwise.
-/// Appends `(S scanned)` when not all scanned files were processed (e.g. where-filters).
-const TAG_MUTATION_FILTER: &str = r#""\(if .dry_run then "[dry-run] " else "" end)\(.tag): \(.modified | length)/\(.total) modified\(if .scanned != .total then " (\(.scanned) scanned)" else "" end)\(if (.modified | length) > 0 then "\n\(.modified | map("  \"\(.)\"") | join("\n"))" else "" end)""#;
-
-/// `BacklinksResult`: `{file, backlinks: [...]}`
-/// Format: `N backlink(s) for "file"` with each backlink listed as `  source.md: line N`.
-/// Empty case: `No backlinks found for "file"`.
-const BACKLINKS_RESULT_FILTER: &str = r#"if (.backlinks | length) == 0 then "No backlinks found for \"\(.file)\"" else "\(.backlinks | length) \(if (.backlinks | length) == 1 then "backlink" else "backlinks" end) for \"\(.file)\"\n\(.backlinks | map("  \(.source): line \(.line)") | join("\n"))" end"#;
-
-/// `LinksFix result`: `{ambiguous, ambiguous_links, applied, applied_fixes, broken, broken_anchors, case_mismatch_fixes, case_mismatches, failed, failed_fixes, fixable, fixes, ignored, relocation_fixes, relocations, unapplied, unapplied_fixes, unfixable, unfixable_links}`
-/// Format: summary line with fix status. Includes case-mismatch, relocation, and ambiguous counts when non-zero.
-/// On `--apply`, the per-fix detail lines show only fixes that were actually
-/// written to disk (`applied_fixes`); on dry-run they show the full plan
-/// (`fixes`), since nothing has been attempted yet. A non-zero `unapplied`
-/// count (fixes whose on-disk text no longer matched the plan) gets its own
-/// section so a stale or partially-applied run is never silently reported as
-/// fully applied.
-///
-/// iter-210 (UX-4) changed three things about the layout:
-/// 1. the `fuzzy` bucket has a **count** line next to `Fixable`/`Unfixable`, so
-///    the summary block accounts for every broken link. Without it a large
-///    vault showed `6098 broken` over `25 fixable + 1400 unfixable` and left
-///    the reader to guess where the other 4,673 went.
-/// 2. `unfixable_links` and `out_of_vault_links` — previously JSON-only — are
-///    listed in text too, capped at 20 entries with an "and N more" footer so
-///    the actionable buckets stay readable on a vault with thousands.
-/// 3. the fuzzy per-fix listing moved to the **end** of the report. It is the
-///    longest section by far and used to bury every actionable bucket under it.
-///
-/// NEW-13 (dogfood pre3) adds a `Relocations` count/section, separate from
-/// `Case mismatches`: a bare-stem link resolved to a different directory is a
-/// move, not a casing fix, and had been silently folded into the case-mismatch
-/// count.
-const LINKS_FIX_FILTER: &str = r#""Broken links: \(.broken)\(if .broken_anchors > 0 then "\n\(.broken_anchors) broken anchor(s) — see `find --broken-links`" else "" end)\nFixable: \(.fixable)\(if .fuzzy > 0 then "\nLow-confidence matches (excluded from plain --apply): \(.fuzzy)" else "" end)\nUnfixable: \(.unfixable)\nIgnored: \(.ignored)\(if .case_mismatches > 0 then "\nCase mismatches: \(.case_mismatches)" else "" end)\(if .relocations > 0 then "\nRelocations: \(.relocations)" else "" end)\(if .ambiguous > 0 then "\nAmbiguous (short-form): \(.ambiguous)" else "" end)\(if .out_of_vault > 0 then "\nOut of vault (target above vault root): \(.out_of_vault)" else "" end)\(if .templated > 0 then "\nTemplated (dynamic destination, never rewritten): \(.templated)" else "" end)\(if .failed > 0 then "\nFailed (write error): \(.failed)\n\(.failed_fixes | map("  \(.source) line \(.line): \"\(.old_target)\" → \"\(.new_target)\" [\(.error)]") | join("\n"))" else "" end)\nApplied: \(if .applied then "yes (\(.applied_fixes | length) fix\(if (.applied_fixes | length) == 1 then "" else "es" end))" else if .dry_run then "no (dry run)" else "no (no fixes written — nothing to apply)" end end)\(if .applied then "\(if (.applied_fixes | length) > 0 then "\n\(.applied_fixes | map("  \(.source) line \(.line): \"\(.old_target)\" → \"\(.new_target)\"") | join("\n"))" else "" end)\(if .unapplied > 0 then "\nUnapplied (plan did not match on-disk text): \(.unapplied)\n\(.unapplied_fixes | map("  \(.source) line \(.line): \"\(.old_target)\" → \"\(.new_target)\"") | join("\n"))" else "" end)" else "\(if (.fixes | length) > 0 then "\n\(.fixes | map("  \(.source) line \(.line): \"\(.old_target)\" → \"\(.new_target)\"") | join("\n"))" else "" end)" end)\(if (.unfixable_links | length) > 0 then "\nUnfixable links (no candidate in the vault):\n\(.unfixable_links | .[:20] | map("  \(.source) line \(.line): \"\(.target)\"") | join("\n"))\(if (.unfixable_links | length) > 20 then "\n  … and \((.unfixable_links | length) - 20) more (use --format json for the full list)" else "" end)" else "" end)\(if (.out_of_vault_links | length) > 0 then "\nOut-of-vault links (target above vault root, never rewritten):\n\(.out_of_vault_links | .[:20] | map("  \(.source) line \(.line): \"\(.target)\"") | join("\n"))\(if (.out_of_vault_links | length) > 20 then "\n  … and \((.out_of_vault_links | length) - 20) more (use --format json for the full list)" else "" end)" else "" end)\(if (.case_mismatch_fixes | length) > 0 then "\nCase-mismatch fixes:\n\(.case_mismatch_fixes | map("  \(.source) line \(.line): \"\(.old_target)\" → \"\(.new_target)\" [\(.rule // "link-case-mismatch")]") | join("\n"))" else "" end)\(if (.relocation_fixes | length) > 0 then "\nRelocation fixes:\n\(.relocation_fixes | map("  \(.source) line \(.line): \"\(.old_target)\" → \"\(.new_target)\" [\(.rule // "shortest-path")]") | join("\n"))" else "" end)\(if (.ambiguous_links | length) > 0 then "\nAmbiguous links:\n\(.ambiguous_links | map("  \(.source) line \(.line): \"\(.target)\" [ambiguous]") | join("\n"))" else "" end)\(if (.templated_links | length) > 0 then "\nTemplated links (dynamic destination, never rewritten):\n\(.templated_links | map("  \(.source) line \(.line): \"\(.target)\" [templated]") | join("\n"))" else "" end)\(if (.fuzzy_fixes | length) > 0 then "\nLow-confidence matches (\(if .fuzzy_applied then "applied at or above confidence \(.fuzzy_min_confidence)" else "not applied — pass --apply-fuzzy" end)):\n\(.fuzzy_fixes | map("  \(.source) line \(.line): \"\(.old_target)\" → \"\(.new_target)\" [\(.rule // "fuzzy-match") \((.confidence * 1000 | floor) / 1000)]\(if .below_floor then " — below floor" else "" end)") | join("\n"))\(if .fuzzy_below_floor > 0 then "\n  \(.fuzzy_below_floor) of \(.fuzzy_fixes | length) below the confidence floor \(.fuzzy_min_confidence) — raise or lower it with --min-confidence <0.0-1.0>" else "" end)" else "" end)""#;
-
-/// `LinksAuto result`: `{ambiguous_titles, applied, apply_outcomes, files_applied, files_failed, files_skipped, matches, scanned, total}`
-/// plus optional `config_excluded_titles` / `config_excluded_mentions`
-/// (iter-195a, renamed and paired in iter-213), present only when
-/// `[links.auto]` config exclusions removed candidate titles — hence the
-/// `// 0` fallbacks in the filter.
-/// Format: summary line + per-match details.
-const LINKS_AUTO_FILTER: &str = r#""\(.matched) unlinked mention\(if .matched == 1 then "" else "s" end) found in \(.matches | map(.file) | unique | length) file\(if (.matches | map(.file) | unique | length) == 1 then "" else "s" end) (\(.scanned) scanned)\(if (.ambiguous_titles | length) > 0 then " (\(.ambiguous_titles | length) ambiguous title\(if (.ambiguous_titles | length) == 1 then "" else "s" end) skipped)" else "" end)\(if (.config_excluded_titles // 0) > 0 then "\nExcluded by [links.auto] config: \(.config_excluded_titles) title\(if .config_excluded_titles == 1 then "" else "s" end), suppressing \(.config_excluded_mentions // 0) mention\(if (.config_excluded_mentions // 0) == 1 then "" else "s" end)" else "" end)\nApplied: \(if .applied then "yes" else "no" end)\(if (.files_failed + .files_skipped) > 0 then "\nWrites: \(.files_applied) applied, \(.files_skipped) skipped, \(.files_failed) failed" else "" end)\(if (.matches | length) > 0 then "\n\(.matches | map("  \(.file):\(.line)    \"\(.matched_text)\" → [[\(.link_target)\(if .matched_text == .link_target then "" else "|\(.matched_text)" end)]]") | join("\n"))" else "" end)""#;
-
-/// `MvResult`: `{dry_run, from, to, total_files_updated, total_links_updated, updated_files}`
-/// Format: `[dry-run] Moved <from> → <to>` with list of updated files and replacements.
-const MV_RESULT_FILTER: &str = r#""\(if .dry_run then "[dry-run] " else "" end)Moved \(.from) → \(.to)\(.updated_files | if length > 0 then "\n" + (map("  \(.file): " + (.replacements | map(.old_text + " → " + .new_text) | join(", "))) | join("\n")) else "" end)""#;
-
-/// `ViewsListEntry`: `{filters, name}`
-/// Format: `name  key=value key=value ...` — compact one-line summary of the view and its filters.
-const VIEWS_LIST_ENTRY_FILTER: &str = r#""\(.name)\t\(.filters | to_entries | map("\(.key)=\(.value | if type == "array" then join(",") else tostring end)") | join(" "))""#;
-
-/// `ViewsMutationResult`: `{action, name}`
-/// Format: `action: name`
-const VIEWS_MUTATION_RESULT_FILTER: &str = r#""\(.action): \(.name)""#;
-
-// ---------------------------------------------------------------------------
-// Shape-based filter lookup
-// ---------------------------------------------------------------------------
-
-/// Compute a sorted comma-joined key signature from a JSON object's top-level keys.
-fn key_signature(map: &serde_json::Map<String, serde_json::Value>) -> String {
-    let mut keys: Vec<&str> = map.keys().map(String::as_str).collect();
-    keys.sort_unstable();
-    keys.join(",")
-}
-
-/// Look up the jq filter for a given key signature.
-///
-/// Returns `None` for unknown shapes, which will fall back to generic formatting.
-fn lookup_filter(key_sig: &str) -> Option<&'static str> {
-    match key_sig {
-        // PropertyInfo
-        "name,type,value" => Some(PROPERTY_INFO_FILTER),
-        // PropertySummaryEntry (mixed_types is skipped when None, so two signatures)
-        "count,name,type" | "count,mixed_types,name,type" => Some(PROPERTY_SUMMARY_ENTRY_FILTER),
-        // TagSummary
-        "tags,total" => Some(TAG_SUMMARY_FILTER),
-        // TagSummaryEntry
-        "count,name" => Some(TAG_SUMMARY_ENTRY_FILTER),
-        // LinkInfo variants (optional path and label → 4 combos).
-        // iter-215 added the always-present `line`; the pre-215 signatures stay
-        // listed so a `LinkInfo` deserialized from an older snapshot or a
-        // hand-written fixture still renders (the filters fall back to
-        // `line 0`) rather than dropping to generic key/value formatting.
-        "line,target" | "target" => Some(LINK_INFO_TARGET_FILTER),
-        "line,path,target" | "path,target" => Some(LINK_INFO_PATH_FILTER),
-        "label,line,target" | "label,target" => Some(LINK_INFO_LABEL_FILTER),
-        "label,line,path,target" | "label,path,target" => Some(LINK_INFO_FULL_FILTER),
-        // LinkInfo carrying a `#fragment` (L-21): fragment is present, plus any
-        // combination of optional path/label, and broken_anchor only when true
-        // (which implies path+fragment). One unified filter handles them all.
-        "fragment,line,target"
-        | "fragment,line,path,target"
-        | "fragment,label,line,target"
-        | "fragment,label,line,path,target"
-        | "broken_anchor,fragment,line,path,target"
-        | "broken_anchor,fragment,label,line,path,target"
-        | "fragment,target"
-        | "fragment,path,target"
-        | "fragment,label,target"
-        | "fragment,label,path,target"
-        | "broken_anchor,fragment,path,target"
-        | "broken_anchor,fragment,label,path,target" => Some(LINK_INFO_ANCHORED_FILTER),
-        // LinkInfo whose target escapes the vault root (iter-193). `path` is
-        // never present; `fragment` / `label` are optional.
-        "line,out_of_vault,target"
-        | "label,line,out_of_vault,target"
-        | "fragment,line,out_of_vault,target"
-        | "fragment,label,line,out_of_vault,target"
-        | "out_of_vault,target"
-        | "label,out_of_vault,target"
-        | "fragment,out_of_vault,target"
-        | "fragment,label,out_of_vault,target" => Some(LINK_INFO_OUT_OF_VAULT_FILTER),
-        // TaskCount
-        "done,total" => Some(TASK_COUNT_FILTER),
-        // OutlineSection (with and without tasks)
-        "code_blocks,heading,level,line,links" => Some(OUTLINE_SECTION_FILTER),
-        "code_blocks,heading,level,line,links,tasks" => Some(OUTLINE_SECTION_WITH_TASKS_FILTER),
-        // TaskInfo
-        "done,line,status,text" => Some(TASK_INFO_FILTER),
-        // FindTaskInfo
-        "done,line,section,status,text" => Some(FIND_TASK_INFO_FILTER),
-        // ContentMatch
-        "line,section,text" => Some(CONTENT_MATCH_FILTER),
-        // TaskReadResult
-        "done,file,line,status,text" => Some(TASK_READ_RESULT_FILTER),
-        // TaskDryRunResult
-        "done,file,line,old_status,status,text" => Some(TASK_DRY_RUN_RESULT_FILTER),
-        // VaultSummary
-        "dead_ends,dir,files,links,orphans,properties,recent_files,status,tags,tasks"
-        | "dead_ends,dir,files,links,orphans,properties,recent_files,schema,status,tags,tasks" => {
-            Some(VAULT_SUMMARY_FILTER)
-        }
-        // Mutation results with property + value (SetPropertyResult, AppendPropertyResult,
-        // RemovePropertyResult with value) — with or without optional `note` field
-        // (iter-216 D-1 added `skipped_count` to all three shapes.)
-        "dry_run,modified,property,scanned,skipped,skipped_count,total,value"
-        | "dry_run,modified,note,property,scanned,skipped,skipped_count,total,value" => {
-            Some(PROPERTY_VALUE_MUTATION_FILTER)
-        }
-        // Mutation results with property only (RemovePropertyResult without value)
-        "dry_run,modified,property,scanned,skipped,skipped_count,total" => {
-            Some(PROPERTY_MUTATION_FILTER)
-        }
-        // Mutation results with tag (SetTagResult, RemoveTagResult)
-        "dry_run,modified,scanned,skipped,skipped_count,tag,total" => Some(TAG_MUTATION_FILTER),
-        // BacklinksResult
-        "backlinks,file" => Some(BACKLINKS_RESULT_FILTER),
-        // LinksFix result (iter-187 adds `failed`/`failed_fixes` for L-11)
-        "ambiguous,ambiguous_links,applied,applied_fixes,broken,broken_anchors,case_mismatch_fixes,case_mismatches,dry_run,failed,failed_fixes,fixable,fixes,fuzzy,fuzzy_applied,fuzzy_below_floor,fuzzy_fixes,fuzzy_min_confidence,ignored,out_of_vault,out_of_vault_links,relocation_fixes,relocations,templated,templated_links,unapplied,unapplied_fixes,unfixable,unfixable_links" => {
-            Some(LINKS_FIX_FILTER)
-        }
-        // LinksAuto result (iter-187 adds per-file apply outcome fields for L-11;
-        // iter-195a adds the config-exclusion attribution, present only when
-        // `[links.auto]` config exclusions removed candidates — hence two
-        // signatures. iter-213 split it into a title count and a mention count,
-        // which always appear together).
-        "ambiguous_titles,applied,apply_outcomes,dry_run,files_applied,files_failed,files_skipped,matched,matches,scanned"
-        | "ambiguous_titles,applied,apply_outcomes,config_excluded_mentions,config_excluded_titles,dry_run,files_applied,files_failed,files_skipped,matched,matches,scanned" => {
-            Some(LINKS_AUTO_FILTER)
-        }
-        // MvResult
-        "dry_run,from,to,total_files_updated,total_links_updated,updated_files" => {
-            Some(MV_RESULT_FILTER)
-        }
-        // ViewsListEntry
-        "filters,name" => Some(VIEWS_LIST_ENTRY_FILTER),
-        // ViewsMutationResult
-        "action,name" => Some(VIEWS_MUTATION_RESULT_FILTER),
-        _ => None,
-    }
-}
-
-// ---------------------------------------------------------------------------
-// jq filter execution engine
-// ---------------------------------------------------------------------------
-
-/// Apply a jq filter string to a `serde_json::Value` and return the text output.
-///
-/// Looks up or compiles the filter in `cache`. Multiple outputs are joined with
-/// newlines. On any error (parse or runtime), returns `None` (used internally
-/// by the text formatter, which has its own fallbacks).
-fn apply_jq_filter(
-    filter_code: &str,
-    value: &serde_json::Value,
-    cache: &mut JaqFilterCache,
-) -> Option<String> {
-    run_jq_filter_cached(filter_code, value, cache).ok()
-}
-
-/// Wall-clock deadline for evaluating a single user-supplied `--jq` filter
-/// (F3-1, deep-analysis-3-2026-08-23.md).
-///
-/// `jaq`'s public API (checked jaq-core 3.0.0) has no step counter, fuel
-/// limit, or other cooperative-cancellation hook — a pathological filter's
-/// unbounded work can happen entirely *inside one* internal evaluation step,
-/// with no opportunity for us to check anything in between. `[range(3e8)]`
-/// builds its whole 300M-element intermediate array before the interpreter
-/// ever yields a value back to Rust (verified: 8.7s / 4.8 GB peak RSS to
-/// print a single number), and `def f: f; f` recurses forever without ever
-/// producing a value at all — so neither is reachable by checking a clock
-/// between values pulled from the output iterator inside
-/// [`execute_jq_filter`]'s loop; that loop's body never even runs for the
-/// second case.
-///
-/// The mitigation is therefore necessarily coarse: run the filter on its own
-/// thread and bound how long the *caller* waits via `recv_timeout`. Both
-/// call sites (`output_pipeline.rs`, `run.rs`'s `hyalo config --jq`) format
-/// the returned error and return almost immediately afterward, so — since
-/// the worker thread is deliberately never joined on timeout — the whole
-/// process, worker included, is torn down by the OS shortly after this
-/// function returns. That caps the *actual* resource exposure to roughly one
-/// deadline's worth of runaway work rather than the filter's full
-/// pathological cost: a 3s deadline turns `[range(3e8)]`'s real 8.7s/4.8 GB
-/// into an error after 3s and however much the abandoned thread allocated in
-/// that window, not the full amount.
-///
-/// A single pathological *value* (not a huge intermediate collection) is
-/// additionally checked by raw byte length before it is ever copied out of
-/// the interpreter — see the pre-check in [`execute_jq_filter`] — since that
-/// case (`"x" * 2000000000`, ~4.0 GB peak RSS in ~1.5s, comfortably under
-/// this deadline) would otherwise slip past both output caps by finishing
-/// before either check could catch it, and used to be measured only *after*
-/// being duplicated into a second multi-GB copy just to learn it was too
-/// big.
-///
-/// **Known residual gap, not covered by anything here:** unbounded *native
-/// stack* recursion — `def f: [f]; f` — overflows the OS thread stack and
-/// hits Rust's SIGSEGV-to-`abort()` guard page, killing the whole process
-/// immediately (verified: `exit 134`, well under this deadline) rather than
-/// erroring cleanly through it. This is not a regression from this
-/// mitigation — the same filter would abort a non-threaded evaluator's
-/// stack identically on unpatched jaq — and there is no user-space hook to
-/// catch a native stack overflow before it happens. Documented rather than
-/// silently claimed to be covered; see DEC-093.
-const JQ_TIME_LIMIT: std::time::Duration = std::time::Duration::from_secs(3);
-
-/// Maximum number of output values a `--jq` filter may emit, independent of
-/// [`JQ_OUTPUT_CAP`]'s byte total.
-///
-/// The byte cap alone only bounds emitted *bytes* — a filter producing
-/// millions of tiny values (e.g. `range(1e9) | tostring[0:0]`, all empty
-/// strings) would iterate far longer than intended without ever crossing
-/// it. This is a cheap, exact check inside the loop that already tracks
-/// `total_len`, catching that class before the byte cap would.
-const JQ_MAX_OUTPUT_VALUES: usize = 1_000_000;
-
-/// Apply a user-supplied jq filter to a `serde_json::Value`.
-///
-/// Compiles the filter on every call. For repeated use across many values,
-/// prefer the cached path via [`format_success`] / [`format_value_as_text`].
-///
-/// Bounded by [`JQ_TIME_LIMIT`]: compilation and execution both run on a
-/// worker thread, and this function returns a clean timeout error if that
-/// thread hasn't produced a result within the deadline (see the constant's
-/// doc comment for why a thread is necessary rather than a cooperative
-/// check). Only `filter_code`/`value` (cloned into the thread) and the final
-/// `Result<String, String>` (sent back over a channel) ever cross the thread
-/// boundary — never a compiled `Filter` or a `jaq_json::Val`, both of which
-/// use `Rc` internally and so are not `Send`.
-///
-/// Returns `Ok(String)` with newline-joined output values on success, or
-/// `Err(String)` with a human-readable description of the parse, runtime, or
-/// timeout error.
-pub fn apply_jq_filter_result(
-    filter_code: &str,
-    value: &serde_json::Value,
-) -> Result<String, String> {
-    let filter_code = filter_code.to_owned();
-    let value = value.clone();
-    let (tx, rx) = std::sync::mpsc::channel();
-
-    let spawned = std::thread::Builder::new()
-        .name("hyalo-jq-eval".to_owned())
-        .spawn(move || {
-            let result = compile_jq_filter(&filter_code)
-                .and_then(|filter| execute_jq_filter(&filter, &value, &filter_code));
-            // The receiver may already have timed out and moved on; a failed
-            // send just means nobody is listening anymore, not an error here.
-            let _ = tx.send(result);
-        });
-    let handle = match spawned {
-        Ok(h) => h,
-        Err(e) => return Err(format!("failed to start jq evaluation: {e}")),
-    };
-
-    match rx.recv_timeout(JQ_TIME_LIMIT) {
-        Ok(result) => {
-            // Finished within the deadline — join is a formality (near-zero
-            // wait, the thread already sent its result), just to avoid
-            // leaking a completed-but-unjoined handle.
-            let _ = handle.join();
-            result
-        }
-        Err(std::sync::mpsc::RecvTimeoutError::Timeout) => {
-            // Deliberately not joined: the worker may still be spinning
-            // (`def f: f; f`) or mid-allocation (`[range(3e8)]`). It is left
-            // to run to completion or to be torn down by the OS when this
-            // process exits shortly after returning this error — see
-            // `JQ_TIME_LIMIT`'s doc comment.
-            drop(handle);
-            Err(format!(
-                "jq filter exceeded the {}s time limit (see `hyalo find --help` for --jq's limits)",
-                JQ_TIME_LIMIT.as_secs()
-            ))
-        }
-        Err(std::sync::mpsc::RecvTimeoutError::Disconnected) => {
-            Err("jq evaluation thread panicked".to_owned())
-        }
-    }
-}
-
-/// Format a jaq load error (lex/parse/IO) into a human-readable string.
-///
-/// `load::Error<&str>` does not implement `Display`, so we extract the first
-/// error's kind and the offending source snippet manually.
-fn format_load_errors(errs: &load::Errors<&str, ()>) -> String {
-    // errs is Vec<(File<&str, ()>, load::Error<&str>)>
-    // We take the first entry and describe its error kind.
-    for (_file, err) in errs {
-        match err {
-            load::Error::Io(ios) => {
-                if let Some((_path, msg)) = ios.first() {
-                    return format!("jq filter error (IO): {msg}");
-                }
-            }
-            load::Error::Lex(lex_errs) => {
-                if let Some((expect, span)) = lex_errs.first() {
-                    return format!(
-                        "jq filter syntax error: expected {} near {:?}",
-                        expect.as_str(),
-                        span
-                    );
-                }
-            }
-            load::Error::Parse(parse_errs) => {
-                if let Some((expect, _token)) = parse_errs.first() {
-                    return format!("jq filter parse error: expected {}", expect.as_str());
-                }
-            }
-        }
-    }
-    "jq filter error: invalid filter syntax".to_owned()
-}
-
-/// Compile a jq filter string into a reusable `Filter`.
-///
-/// The `Arena` used during loading is a temporary scratch pad and is dropped
-/// after this function returns — the compiled `Filter` owns all its data.
-fn compile_jq_filter(filter_code: &str) -> Result<jaq_core::compile::Filter<Native<D>>, String> {
-    let program = File {
-        code: filter_code,
-        path: (),
-    };
-    let defs = jaq_core::defs()
-        .chain(jaq_std::defs())
-        .chain(jaq_json::defs());
-    let loader = Loader::new(defs);
-    let arena = Arena::default();
-
-    let modules = loader
-        .load(&arena, program)
-        .map_err(|errs| format_load_errors(&errs))?;
-
-    let funs = jaq_core::funs::<D>()
-        .chain(jaq_std::funs::<D>())
-        .chain(jaq_json::funs::<D>());
-    Compiler::default()
-        .with_funs(funs)
-        .compile(modules)
-        .map_err(|errs| {
-            // compile::Errors = Vec<(File<S,P>, Vec<(S, Undefined)>)>
-            // Extract the first undefined symbol name for a useful message.
-            let first = errs.iter().flat_map(|(_file, undefs)| undefs.iter()).next();
-            if let Some((name, undef)) = first {
-                format!("jq filter error: undefined {} {:?}", undef.as_str(), name)
-            } else {
-                "jq filter error: compilation failed".to_owned()
-            }
-        })
-}
-
-/// Maximum total output size for a jq filter to prevent pathological filters
-/// from causing unbounded memory growth (e.g. exponential-expansion patterns).
-const JQ_OUTPUT_CAP: usize = 10 * 1024 * 1024; // 10 MiB
-
-/// Maximum length (in `char`s) of a diagnostic string embedded in an error
-/// envelope — the jq runtime error message and the filter code that produced
-/// it (F-3, `deep-analysis-2-2026-08-23.md`). jaq's runtime `Display` embeds
-/// the *value* it failed on (e.g. the whole `.results` array for `.results |
-/// .file` on an array), which on a large vault can be megabytes; an
-/// unbounded error also makes `--jq` a content-disclosure vector for any
-/// consumer that logs error output. This is a `char` bound, not a byte
-/// bound, so truncation never splits a multi-byte codepoint.
-const JQ_ERROR_DIAGNOSTIC_CHAR_CAP: usize = 200;
-
-/// Truncate `s` to at most [`JQ_ERROR_DIAGNOSTIC_CHAR_CAP`] characters,
-/// appending `…` when truncation occurred. Character-based (not byte-based)
-/// so a multi-byte codepoint is never split.
-fn truncate_diagnostic(s: &str) -> String {
-    if s.chars().count() <= JQ_ERROR_DIAGNOSTIC_CHAR_CAP {
-        return s.to_owned();
-    }
-    let mut truncated: String = s.chars().take(JQ_ERROR_DIAGNOSTIC_CHAR_CAP).collect();
-    truncated.push('…');
-    truncated
-}
-
-/// Execute a pre-compiled jq filter against a JSON value and return the text output.
-///
-/// `filter_code` is used only to name the filter in error messages — the
-/// filter itself has already been compiled into `filter`.
-fn execute_jq_filter(
-    filter: &jaq_core::compile::Filter<Native<D>>,
-    value: &serde_json::Value,
-    filter_code: &str,
-) -> Result<String, String> {
-    let input: Val = serde_json::from_value(value.clone()).map_err(|e| {
-        format!(
-            "jq input conversion error in filter {}: {}",
-            truncate_diagnostic(filter_code),
-            truncate_diagnostic(&e.to_string())
-        )
-    })?;
-    let ctx = Ctx::<D>::new(&filter.lut, Vars::new([]));
-
-    let mut out = String::new();
-    let mut total_len: usize = 0;
-    let mut value_count: usize = 0;
-    for result in filter.id.run((ctx, input)).map(jaq_core::unwrap_valr) {
-        match result {
-            Ok(val) => {
-                value_count += 1;
-                if value_count > JQ_MAX_OUTPUT_VALUES {
-                    return Err(format!(
-                        "jq filter output exceeds {JQ_MAX_OUTPUT_VALUES} values"
-                    ));
-                }
-                // Check a string value's raw byte length BEFORE copying it
-                // (Finding 2, review round on PR #254): a single pathological
-                // value — `"x" * 2000000000` measured at ~4.0 GB peak RSS in
-                // 1.49s, comfortably under JQ_TIME_LIMIT — used to be
-                // measured only *after* `.to_owned()`/`from_utf8_lossy()`
-                // duplicated it into a second multi-GB buffer just to learn
-                // it was too big. `s` here still borrows from `val`, so this
-                // costs nothing beyond the length check already available on
-                // the byte slice jaq handed us.
-                //
-                // No equivalent pre-check exists for non-string values
-                // (`other.to_string()` below): jaq's `Display` impl builds
-                // the whole formatted string in one call with no length
-                // preview, so a single huge *non-string* value (e.g. a large
-                // array/object formatted directly, without a `length`/count
-                // reduction) is bounded only by JQ_TIME_LIMIT, not by either
-                // output cap — documented on `JQ_TIME_LIMIT`'s doc comment
-                // and in DEC-093 rather than silently claimed to be covered.
-                if let Val::TStr(ref s) | Val::BStr(ref s) = val
-                    && s.len() > JQ_OUTPUT_CAP
-                {
-                    return Err(format!(
-                        "jq filter output exceeds {} MiB limit",
-                        JQ_OUTPUT_CAP / (1024 * 1024)
-                    ));
-                }
-                let s = match val {
-                    Val::TStr(ref s) | Val::BStr(ref s) => match std::str::from_utf8(s) {
-                        Ok(valid) => valid.to_owned(),
-                        Err(_) => String::from_utf8_lossy(s).into_owned(),
-                    },
-                    // For non-string values, `Display` produces valid JSON
-                    // (numbers, booleans, null, arrays, objects).
-                    other => other.to_string(),
-                };
-                // Account for the newline separator that will be prepended
-                // between fragments when out is non-empty.
-                total_len = total_len
-                    .saturating_add(s.len())
-                    .saturating_add(usize::from(!out.is_empty()));
-                if total_len > JQ_OUTPUT_CAP {
-                    return Err(format!(
-                        "jq filter output exceeds {} MiB limit",
-                        JQ_OUTPUT_CAP / (1024 * 1024)
-                    ));
-                }
-                if !out.is_empty() {
-                    out.push('\n');
-                }
-                out.push_str(&s);
-            }
-            Err(e) => {
-                // jaq's runtime `Display` embeds the value it failed on (e.g. the
-                // whole input array for `.file` applied to an array) — truncate
-                // both the value-bearing message and the filter code so a single
-                // mistyped `--jq` filter on a large vault can't dump megabytes of
-                // vault content into the error envelope (F-3, DEC-094).
-                return Err(format!(
-                    "jq runtime error in filter {}: {}",
-                    truncate_diagnostic(filter_code),
-                    truncate_diagnostic(&e.to_string())
-                ));
-            }
-        }
-    }
-
-    Ok(out)
-}
-
-/// Look up or compile a jq filter from `cache`, then execute it against `value`.
-fn run_jq_filter_cached(
-    filter_code: &str,
-    value: &serde_json::Value,
-    cache: &mut JaqFilterCache,
-) -> Result<String, String> {
-    if let Some(filter) = cache.get(filter_code) {
-        return execute_jq_filter(filter, value, filter_code);
-    }
-    let compiled = compile_jq_filter(filter_code)?;
-    let filter = cache.entry(filter_code.to_owned()).or_insert(compiled);
-    execute_jq_filter(filter, value, filter_code)
-}
-
-// ---------------------------------------------------------------------------
-// FileObject dynamic filter builder
-// ---------------------------------------------------------------------------
-
-/// Build a jaq filter string for a `FileObject` by inspecting which optional
-/// fields are present in the JSON object.
-///
-/// The file header is always emitted. Each optional section (properties, tags,
-/// sections, tasks, matches, links) is included only when the key is present.
-///
-/// **How it works:** Each part is a jaq expression that either emits a string or
-/// `empty` (when the field is absent/empty). Parts are joined with `, ` — jaq's
-/// alternation operator — so the filter produces one output per present section.
-/// `run_jq_filter` then joins those outputs with `"\n"`, producing the final
-/// multi-line text block. This coupling is intentional: changing the separator
-/// in `run_jq_filter` would affect `FileObject` rendering.
-fn build_file_object_filter(map: &serde_json::Map<String, serde_json::Value>) -> String {
-    // Header: file path and modified timestamp — always present.
-    let mut parts = vec![r#""\"\(.file)\"  (\(.modified))""#.to_owned()];
-
-    // Title: "  title: <value>" or "  title: (none)"
-    if map.contains_key("title") {
-        parts.push(r#""  title: \(if .title != null then .title else "(none)" end)""#.to_owned());
-    }
-
-    // Properties: header then each as "    key: value"
-    if map.contains_key("properties") {
-        parts.push(
-            r#"if (.properties | length) > 0 then "  properties:\n\(.properties | to_entries | map("    \(.key): \(if (.value | type) == "array" then "[" + (.value | map(tostring) | join(", ")) + "]" else .value end)") | join("\n"))" else empty end"#.to_owned(),
-        );
-    }
-
-    // Properties (typed): header then each as "    name (type): value"
-    if map.contains_key("properties_typed") {
-        parts.push(
-            r#"if (.properties_typed | length) > 0 then "  properties_typed:\n\(.properties_typed | map("    \(.name) (\(.type)): \(if (.value | type) == "array" then "[" + (.value | map(tostring) | join(", ")) + "]" else .value end)") | join("\n"))" else empty end"#.to_owned(),
-        );
-    }
-
-    // Tags: "  tags: [tag1, tag2, ...]"
-    if map.contains_key("tags") {
-        parts.push(
-            r#"if (.tags | length) > 0 then "  tags: [\(.tags | join(", "))]" else empty end"#
-                .to_owned(),
-        );
-    }
-
-    // Sections: header then each as "    ## Heading [done/total]" or "    ## Heading"
-    // Note: uses r##"..."## because the jq filter contains the sequence "#" (hash-quoted).
-    // NEW-16 (dogfood pre3): when a computed count is about to be appended,
-    // strip a trailing hand-written `[n/m]` from `.heading` first — see
-    // OUTLINE_SECTION_WITH_TASKS_FILTER's doc comment for why. Only stripped
-    // when `.tasks` is present: a heading with no task section keeps its text
-    // exactly as written, even if it happens to end in bracket text that
-    // merely looks like a count.
-    if map.contains_key("sections") {
-        parts.push(
-            r##"if (.sections | length) > 0 then "  sections:\n\(.sections | map("    \("#" * .level) \(if .tasks then ((.heading // "(pre-heading)") | sub("\\s*\\[[0-9]+/[0-9]+\\]\\s*$"; "")) else (.heading // "(pre-heading)") end)\(if .tasks then " [\(.tasks.done)/\(.tasks.total)]" else "" end)") | join("\n"))" else empty end"##.to_owned(),
-        );
-    }
-
-    // Tasks: header then each as "    [x] text (line N)"
-    if map.contains_key("tasks") {
-        parts.push(
-            r#"if (.tasks | length) > 0 then "  tasks:\n\(.tasks | map("    [\(if .done then "x" else " " end)] \(.text) (line \(.line))") | join("\n"))" else empty end"#.to_owned(),
-        );
-    }
-
-    // Matches: header then each as "    line N (section): text"
-    if map.contains_key("matches") {
-        parts.push(
-            r#"if (.matches | length) > 0 then "  matches:\n\(.matches | map("    line \(.line) (\(.section)): \(.text)") | join("\n"))" else empty end"#.to_owned(),
-        );
-    }
-
-    // Score: "  score: <value>" — BM25 relevance score when pattern search was used
-    if map.contains_key("score") {
-        parts.push(r#""  score: \(.score)""#.to_owned());
-    }
-
-    // Links: header then each as "    line N: \"target\" → \"path\"" or
-    // "    line N: \"target\" (unresolved)".
-    // Anchored links (L-21) append "#fragment" to the target and, when the
-    // heading is missing, " (broken anchor)" after the path — keep in sync
-    // with LINK_INFO_ANCHORED_FILTER, which renders the standalone shape.
-    // iter-215 (dogfood UX-6): the `line N:` prefix is what makes a
-    // `find --broken-links` report actionable — it names the line the broken
-    // link is written on, the same one `lint`/`backlinks` report. `// 0`
-    // guards JSON from a pre-215 hyalo that has no `.line`.
-    if map.contains_key("links") {
-        parts.push(
-            r##"if (.links | length) > 0 then "  links:\n\(.links | map("    line \(.line // 0): \"\(.target)\(if .fragment then "#\(.fragment)" else "" end)\"\(if .path then " → \"\(.path)\"\(if .broken_anchor then " (broken anchor)" else "" end)" else (if .out_of_vault then " (out of vault)" else " (unresolved)" end) end)") | join("\n"))" else empty end"##.to_owned(),
-        );
-    }
-
-    // Backlinks: header then each as "    \"source\" line N" or "    \"source\" line N: label"
-    if map.contains_key("backlinks") {
-        parts.push(
-            r#"if (.backlinks | length) > 0 then "  backlinks:\n\(.backlinks | map("    \"\(.source)\" line \(.line)\(if .label then ": \(.label)" else "" end)") | join("\n"))" else empty end"#.to_owned(),
-        );
-    }
-
-    parts.join(", ")
-}
-
-// ---------------------------------------------------------------------------
-// Text formatting
-// ---------------------------------------------------------------------------
-
-/// Format a JSON value as human-readable text using jq filters where available.
-fn format_value_as_text(value: &serde_json::Value, cache: &mut JaqFilterCache) -> String {
-    match value {
-        serde_json::Value::Array(arr) => {
-            // TypeList: array of type list entries — use custom formatter with blank-line separation.
-            let is_type_list = arr.first().and_then(|v| v.as_object()).is_some_and(|m| {
-                key_signature(m) == "has_filename_template,property_count,required,type"
-            });
-            if is_type_list {
-                return arr
-                    .iter()
-                    .filter_map(|v| v.as_object())
-                    .map(format_type_list_entry_text)
-                    .collect::<Vec<_>>()
-                    .join("\n\n");
-            }
-            // LintRules list: array of rule entries with id, effective_enabled, etc.
-            let is_lint_rules = arr.first().and_then(|v| v.as_object()).is_some_and(|m| {
-                m.contains_key("id")
-                    && m.contains_key("effective_enabled")
-                    && m.contains_key("autofixable")
-                    && m.contains_key("source")
-            });
-            if is_lint_rules {
-                let result = format_lint_rules_list_text(arr);
-                if !result.is_empty() {
-                    return result;
-                }
-            }
-            // Use blank-line separator between FileObjects for readability.
-            let is_file_objects = arr
-                .first()
-                .and_then(|v| v.as_object())
-                .is_some_and(|m| m.contains_key("file") && m.contains_key("modified"));
-            let sep = if is_file_objects { "\n\n" } else { "\n" };
-            arr.iter()
-                .map(|v| format_value_as_text(v, cache))
-                .collect::<Vec<_>>()
-                .join(sep)
-        }
-        serde_json::Value::Object(map) => {
-            let sig = key_signature(map);
-            if let Some(filter) = lookup_filter(&sig)
-                && let Some(output) = apply_jq_filter(filter, value, cache)
-            {
-                return output;
-            }
-            // TypeShow: detected by presence of "properties" object + "required" array + "type" string.
-            // Accept both the old signature (without required_sections) and the new one.
-            if sig == "defaults,filename_template,properties,required,type"
-                || sig == "defaults,filename_template,properties,required,required_sections,type"
-            {
-                return format_type_show_text(map);
-            }
-            // LintFixOutput (fix-mode): detected by `total_fixed` + `files`.
-            if map.contains_key("total_fixed")
-                && map.contains_key("files")
-                && map.contains_key("total_remaining")
-            {
-                return format_lint_fix_output_text(map);
-            }
-            // Lint output (`ExtLintOutput`): detected by "files" array of
-            // {file, violations} plus the run-level violation count. That count
-            // was renamed `total` -> `violations` in iter-216 (D-2); `total`
-            // stays in the predicate so a payload produced by an older hyalo
-            // still renders as lint text.
-            if (map.contains_key("violations") || map.contains_key("total"))
-                && map.contains_key("files")
-                && let Some(serde_json::Value::Array(arr)) = map.get("files")
-            {
-                let is_lint = arr
-                    .first()
-                    .and_then(|v| v.as_object())
-                    .is_some_and(|m| {
-                        m.contains_key("file")
-                            && (m.contains_key("violations") || m.contains_key("rule_groups"))
-                    })
-                    // Empty file list with new-shape totals counts as lint.
-                    || (arr.is_empty()
-                        && (map.contains_key("rules_fired")
-                            || map.contains_key("files_checked")));
-                if is_lint {
-                    return format_lint_output_text(map);
-                }
-            }
-            // LintRules mutation (set/remove): detected by `action` = "set" or "remove"
-            // with `rule_id`, `before`, `after`, `config_path` fields.
-            if matches!(
-                map.get("action").and_then(serde_json::Value::as_str),
-                Some("set" | "remove")
-            ) && map.contains_key("rule_id")
-                && map.contains_key("before")
-                && map.contains_key("after")
-                && map.contains_key("config_path")
-            {
-                return format_lint_rules_mutation_text(map);
-            }
-            // FileObject: dynamically compose filter from present fields.
-            if map.contains_key("file") && map.contains_key("modified") {
-                let filter = build_file_object_filter(map);
-                if let Some(output) = apply_jq_filter(&filter, value, cache) {
-                    return output;
-                }
-            }
-            // Generator results (`okf index`, `okf log`, `madr toc`): keyed on the
-            // `command` string. These carry nested `files` arrays / action fields
-            // that the generic key:value dump renders unreadably.
-            if let Some(cmd) = map.get("command").and_then(serde_json::Value::as_str)
-                && let Some(out) = format_generator_output_text(cmd, map)
-            {
-                return out;
-            }
-            // Fallback: generic key: value lines
-            format_object_generic(map, cache)
-        }
-        other => format_scalar(other, cache),
-    }
-}
-
-/// Format `ExtLintOutput` JSON as human-readable text.
-///
-/// Reproduces the format previously generated by `commands::lint::format_text_output`.
-fn format_lint_output_text(map: &serde_json::Map<String, serde_json::Value>) -> String {
-    use std::fmt::Write as _;
-
-    let mut s = String::new();
-    let dry_run = map
-        .get("dry_run")
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or(false);
-
-    // Fix actions (shown first).
-    if let Some(fixes_arr) = map.get("fixes").and_then(|f| f.as_array()) {
-        let verb = if dry_run { "Would fix" } else { "Fixed" };
-        for file_fix in fixes_arr {
-            let file = file_fix
-                .get("file")
-                .and_then(serde_json::Value::as_str)
-                .unwrap_or("?");
-            let actions = file_fix.get("actions").and_then(|a| a.as_array());
-            let Some(actions) = actions else { continue };
-            if actions.is_empty() {
-                continue;
-            }
-            let _ = writeln!(s, "{verb} {file}:");
-            for a in actions {
-                let kind = a
-                    .get("kind")
-                    .and_then(serde_json::Value::as_str)
-                    .unwrap_or("");
-                let property = a
-                    .get("property")
-                    .and_then(serde_json::Value::as_str)
-                    .unwrap_or("?");
-                let new = a
-                    .get("new")
-                    .and_then(serde_json::Value::as_str)
-                    .unwrap_or("");
-                let old = a.get("old").and_then(serde_json::Value::as_str);
-                match (kind, old) {
-                    ("insert-default", _) => {
-                        let _ = writeln!(s, "  insert  {property} = {new:?}");
-                    }
-                    ("infer-type", _) => {
-                        let _ = writeln!(s, "  infer   type = {new:?}");
-                    }
-                    ("fix-enum-typo", Some(old_v)) => {
-                        let _ = writeln!(s, "  enum    {property}: {old_v:?} -> {new:?}");
-                    }
-                    ("normalize-date", Some(old_v)) => {
-                        let _ = writeln!(s, "  date    {property}: {old_v:?} -> {new:?}");
-                    }
-                    _ => {
-                        let _ = writeln!(s, "  {kind}  {property} = {new:?}");
-                    }
-                }
-            }
-        }
-    }
-
-    // File violations — handle both the new `rule_groups` shape and the legacy `violations` shape.
-    let files = map.get("files").and_then(|f| f.as_array());
-    if let Some(files) = files {
-        for file_entry in files {
-            let file = file_entry
-                .get("file")
-                .and_then(serde_json::Value::as_str)
-                .unwrap_or("?");
-
-            // New shape: violations are grouped by rule under `rule_groups`.
-            if let Some(rule_groups) = file_entry.get("rule_groups").and_then(|rg| rg.as_array()) {
-                if rule_groups.is_empty() {
-                    continue;
-                }
-                let _ = writeln!(s, "{file}:");
-                for group in rule_groups {
-                    let rule = group
-                        .get("rule")
-                        .and_then(serde_json::Value::as_str)
-                        .unwrap_or("?");
-                    // Group severity is the fallback when a violation does not
-                    // carry its own (older payloads); a folded group such as
-                    // `SCHEMA` mixes severities, so each line is labelled from
-                    // its own `severity` field to agree with the counts (BUG-17).
-                    let group_severity = group
-                        .get("severity")
-                        .and_then(serde_json::Value::as_str)
-                        .unwrap_or("warn");
-                    let violations = group.get("violations").and_then(|v| v.as_array());
-                    let count = group
-                        .get("count")
-                        .and_then(serde_json::Value::as_u64)
-                        .unwrap_or(0);
-                    let shown = group
-                        .get("shown")
-                        .and_then(serde_json::Value::as_u64)
-                        .unwrap_or(count);
-                    if let Some(violations) = violations {
-                        for v in violations {
-                            let line = v
-                                .get("line")
-                                .and_then(serde_json::Value::as_u64)
-                                .unwrap_or(0);
-                            let message = v
-                                .get("message")
-                                .and_then(serde_json::Value::as_str)
-                                .unwrap_or("");
-                            let severity = v
-                                .get("severity")
-                                .and_then(serde_json::Value::as_str)
-                                .unwrap_or(group_severity);
-                            let pad = if severity == "error" {
-                                "error"
-                            } else {
-                                "warn "
-                            };
-                            if line > 0 {
-                                let _ = writeln!(s, "  {pad}  {rule}  line {line}  {message}");
-                            } else {
-                                let _ = writeln!(s, "  {pad}  {rule}  {message}");
-                            }
-                        }
-                        if count > shown {
-                            let _ = writeln!(s, "  … ({} more {})", count - shown, rule);
-                        }
-                    }
-                }
-            } else if let Some(violations) = file_entry.get("violations").and_then(|v| v.as_array())
-            {
-                // Legacy shape: flat violations array.
-                if violations.is_empty() {
-                    continue;
-                }
-                let _ = writeln!(s, "{file}:");
-                for v in violations {
-                    let severity = v
-                        .get("severity")
-                        .and_then(serde_json::Value::as_str)
-                        .unwrap_or("warn");
-                    let message = v
-                        .get("message")
-                        .and_then(serde_json::Value::as_str)
-                        .unwrap_or("");
-                    let pad = if severity == "error" {
-                        "error"
-                    } else {
-                        "warn "
-                    };
-                    let _ = writeln!(s, "  {pad}  {message}");
-                }
-            }
-        }
-    }
-
-    let error_count: u64 = map
-        .get("errors")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    let warn_count: u64 = map
-        .get("warnings")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    // Support both old `files_with_issues` and new `files_with_violations`.
-    let files_with_issues: u64 = map
-        .get("files_with_violations")
-        .or_else(|| map.get("files_with_issues"))
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    // Support both old `limited` and new `files_truncated`.
-    let limited = map
-        .get("files_truncated")
-        .or_else(|| map.get("limited"))
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or(false);
-    let shown_files = map
-        .get("files")
-        .and_then(|f| f.as_array())
-        .map_or(0, |arr| {
-            arr.iter()
-                .filter(|e| {
-                    // New shape: has rule_groups
-                    e.get("rule_groups")
-                        .and_then(|rg| rg.as_array())
-                        .is_some_and(|rg| !rg.is_empty())
-                        // Legacy shape: has violations
-                        || e.get("violations")
-                            .and_then(|v| v.as_array())
-                            .is_some_and(|v| !v.is_empty())
-                })
-                .count()
-        });
-
-    if limited {
-        let _ = writeln!(
-            s,
-            "… (showing {shown_files} of {files_with_issues} files with issues)"
-        );
-    }
-
-    // Summary line.
-    let files_checked: u64 = map
-        .get("files_checked")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    let files_label = if files_checked == 1 { "file" } else { "files" };
-    if error_count == 0 && warn_count == 0 {
-        let _ = write!(s, "{files_checked} {files_label} checked, no issues");
-    } else {
-        let errors_label = if error_count == 1 { "error" } else { "errors" };
-        let warns_label = if warn_count == 1 {
-            "warning"
-        } else {
-            "warnings"
-        };
-        let _ = write!(
-            s,
-            "{files_checked} {files_label} checked, {files_with_issues} with issues ({error_count} {errors_label}, {warn_count} {warns_label})",
-        );
-    }
-    // UX-1 (dogfood pre3): a bare sweep silently dropped every `[lint]
-    // ignore`-matched file with no visible trace — "68 files checked, no
-    // issues" on a 386-file vault read as a clean bill of health, not "68 of
-    // 386 were actually looked at." The named-file / glob-all-ignored cases
-    // already get their own stderr notice; this is the everything-else case.
-    let files_ignored: u64 = map
-        .get("files_ignored")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    if files_ignored > 0 {
-        let _ = write!(s, " ({files_ignored} ignored by [lint] ignore)");
-    }
-
-    let fix_count: usize = map
-        .get("fixes")
-        .and_then(|f| f.as_array())
-        .map_or(0, |arr| {
-            arr.iter()
-                .filter_map(|f| f.get("actions").and_then(|a| a.as_array()).map(Vec::len))
-                .sum()
-        });
-    if fix_count > 0 {
-        let fixed_label = if dry_run { "would fix" } else { "fixed" };
-        let _ = write!(s, " — {fixed_label} {fix_count}");
-    }
-
-    s
-}
-
-/// Format `ExtLintFixOutput` (fix-mode lint) as human-readable text.
-///
-/// Renders three prefixes: `fixed`, `remain`, `conflict`.
-fn format_lint_fix_output_text(map: &serde_json::Map<String, serde_json::Value>) -> String {
-    use std::fmt::Write as _;
-
-    let mut s = String::new();
-    let dry_run = map
-        .get("dry_run")
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or(false);
-
-    let files = map.get("files").and_then(|f| f.as_array());
-    if let Some(files) = files {
-        for file_entry in files {
-            let file = file_entry
-                .get("file")
-                .and_then(serde_json::Value::as_str)
-                .unwrap_or("?");
-
-            let has_content = file_entry
-                .get("fixed_groups")
-                .and_then(|g| g.as_array())
-                .is_some_and(|a| !a.is_empty())
-                || file_entry
-                    .get("remaining_groups")
-                    .and_then(|g| g.as_array())
-                    .is_some_and(|a| !a.is_empty())
-                || file_entry
-                    .get("conflicts")
-                    .and_then(|c| c.as_array())
-                    .is_some_and(|a| !a.is_empty());
-            if !has_content {
-                continue;
-            }
-
-            let _ = writeln!(s, "{file}:");
-            let prefix = if dry_run { "would fix" } else { "fixed  " };
-
-            // Fixed groups.
-            if let Some(fixed_groups) = file_entry.get("fixed_groups").and_then(|g| g.as_array()) {
-                for group in fixed_groups {
-                    let rule = group
-                        .get("rule")
-                        .and_then(serde_json::Value::as_str)
-                        .unwrap_or("?");
-                    let violations = group.get("violations").and_then(|v| v.as_array());
-                    let count = group
-                        .get("count")
-                        .and_then(serde_json::Value::as_u64)
-                        .unwrap_or(0);
-                    if let Some(violations) = violations.filter(|v| !v.is_empty()) {
-                        for v in violations {
-                            let line = v
-                                .get("line")
-                                .and_then(serde_json::Value::as_u64)
-                                .unwrap_or(0);
-                            let message = v
-                                .get("message")
-                                .and_then(serde_json::Value::as_str)
-                                .unwrap_or("");
-                            if line > 0 {
-                                let _ = writeln!(s, "  {prefix}  {rule}  line {line}  {message}");
-                            } else {
-                                let _ = writeln!(s, "  {prefix}  {rule}  {message}");
-                            }
-                        }
-                    } else if count > 0 {
-                        let _ = writeln!(s, "  {prefix}  {rule}  ({count} violations)");
-                    }
-                }
-            }
-
-            // Remaining groups.
-            if let Some(remaining_groups) = file_entry
-                .get("remaining_groups")
-                .and_then(|g| g.as_array())
-            {
-                for group in remaining_groups {
-                    let rule = group
-                        .get("rule")
-                        .and_then(serde_json::Value::as_str)
-                        .unwrap_or("?");
-                    let severity = group
-                        .get("severity")
-                        .and_then(serde_json::Value::as_str)
-                        .unwrap_or("warn");
-                    let sev_pad = if severity == "error" {
-                        "error"
-                    } else {
-                        "warn "
-                    };
-                    let violations = group.get("violations").and_then(|v| v.as_array());
-                    let count = group
-                        .get("count")
-                        .and_then(serde_json::Value::as_u64)
-                        .unwrap_or(0);
-                    let shown = group
-                        .get("shown")
-                        .and_then(serde_json::Value::as_u64)
-                        .unwrap_or(count);
-                    if let Some(violations) = violations {
-                        for v in violations {
-                            let line = v
-                                .get("line")
-                                .and_then(serde_json::Value::as_u64)
-                                .unwrap_or(0);
-                            let message = v
-                                .get("message")
-                                .and_then(serde_json::Value::as_str)
-                                .unwrap_or("");
-                            if line > 0 {
-                                let _ = writeln!(
-                                    s,
-                                    "  remain  {sev_pad}  {rule}  line {line}  {message}"
-                                );
-                            } else {
-                                let _ = writeln!(s, "  remain  {sev_pad}  {rule}  {message}");
-                            }
-                        }
-                        if count > shown {
-                            let _ = writeln!(s, "  … ({} more {})", count - shown, rule);
-                        }
-                    }
-                }
-            }
-
-            // Conflicts.
-            //
-            // A rule with several violations in one file can have some fixes
-            // applied and one lose a range overlap, which printed the rule as
-            // both `fixed` and `conflict` two lines apart — read as a
-            // contradiction rather than as two different violations (iter-213,
-            // dogfood UX-5). The `fixed` line is the one that describes what
-            // changed on disk, so the conflict line is dropped when the same
-            // rule already appears there. The JSON keeps both: `conflicts` is
-            // how a consumer learns a fix was skipped, and dropping entries
-            // from it would lose that.
-            let fixed_rules: std::collections::HashSet<&str> = file_entry
-                .get("fixed_groups")
-                .and_then(|g| g.as_array())
-                .map(|groups| {
-                    groups
-                        .iter()
-                        .filter_map(|g| g.get("rule").and_then(serde_json::Value::as_str))
-                        .collect()
-                })
-                .unwrap_or_default();
-            if let Some(conflicts) = file_entry.get("conflicts").and_then(|c| c.as_array()) {
-                for conflict in conflicts {
-                    let rule = conflict
-                        .get("rule")
-                        .and_then(serde_json::Value::as_str)
-                        .unwrap_or("?");
-                    if fixed_rules.contains(rule) {
-                        continue;
-                    }
-                    let reason = conflict
-                        .get("reason")
-                        .and_then(serde_json::Value::as_str)
-                        .unwrap_or("");
-                    let _ = writeln!(s, "  conflict  {rule}  {reason}");
-                }
-            }
-        }
-    }
-
-    // `files_truncated` gates a "(showing N of M files with issues)" line,
-    // mirroring `format_lint_output_text`'s. Without it, the summary line
-    // below — whose fixed/remaining/conflicts counts describe the WHOLE run
-    // since iter-218 (NEW-6) — read as if the file listing above it were
-    // complete too, when `--limit` had silently capped it (review finding
-    // #2).
-    let files_truncated = map
-        .get("files_truncated")
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or(false);
-    if files_truncated {
-        let files_with_violations = map
-            .get("files_with_violations")
-            .and_then(serde_json::Value::as_u64)
-            .unwrap_or(0);
-        let shown_files = files.map_or(0, |arr| {
-            arr.iter()
-                .filter(|f| {
-                    f.get("fixed_groups")
-                        .and_then(|g| g.as_array())
-                        .is_some_and(|a| !a.is_empty())
-                        || f.get("remaining_groups")
-                            .and_then(|g| g.as_array())
-                            .is_some_and(|a| !a.is_empty())
-                        || f.get("conflicts")
-                            .and_then(|c| c.as_array())
-                            .is_some_and(|a| !a.is_empty())
-                })
-                .count()
-        });
-        let _ = writeln!(
-            s,
-            "… (showing {shown_files} of {files_with_violations} files with issues)"
-        );
-    }
-
-    // Summary line.
-    let files_checked = map
-        .get("files_checked")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    let total_fixed = map
-        .get("total_fixed")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    let total_remaining = map
-        .get("total_remaining")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    let total_conflicts = map
-        .get("total_conflicts")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    let files_label = if files_checked == 1 { "file" } else { "files" };
-    let fixed_phrase = if dry_run {
-        format!("would fix {total_fixed}")
-    } else {
-        format!("fixed {total_fixed}")
-    };
-    let _ = write!(
-        s,
-        "{files_checked} {files_label} checked: {fixed_phrase} · remaining {total_remaining} · conflicts {total_conflicts}.",
-    );
-
-    s
-}
-
-/// Format a `lint-rules list` result as a column table.
-///
-/// Detects arrays of objects with `id`, `default_severity`, `effective_enabled`,
-/// `autofixable`, `source` fields.
-fn format_lint_rules_list_text(arr: &[serde_json::Value]) -> String {
-    use std::fmt::Write as _;
-
-    const DEFAULT_TERM_WIDTH: usize = 120;
-
-    if arr.is_empty() {
-        return String::new();
-    }
-
-    // Verify it looks like lint-rule entries.
-    let first_is_rule = arr.first().and_then(|v| v.as_object()).is_some_and(|m| {
-        m.contains_key("id")
-            && m.contains_key("effective_enabled")
-            && m.contains_key("autofixable")
-            && m.contains_key("source")
-    });
-    if !first_is_rule {
-        return String::new();
-    }
-
-    let term_width = DEFAULT_TERM_WIDTH;
-
-    // Collect rows.
-    let mut rows: Vec<[String; 6]> = Vec::with_capacity(arr.len());
-    let mut total_enabled = 0usize;
-    let mut total_autofixable = 0usize;
-
-    for entry in arr {
-        let Some(obj) = entry.as_object() else {
-            continue;
-        };
-        let id = obj
-            .get("id")
-            .and_then(serde_json::Value::as_str)
-            .unwrap_or("")
-            .to_owned();
-        let name = obj
-            .get("name")
-            .and_then(serde_json::Value::as_str)
-            .unwrap_or("")
-            .to_owned();
-        let severity = obj
-            .get("effective_severity")
-            .or_else(|| obj.get("default_severity"))
-            .and_then(serde_json::Value::as_str)
-            .unwrap_or("warn")
-            .to_owned();
-        let enabled = obj
-            .get("effective_enabled")
-            .and_then(serde_json::Value::as_bool)
-            .unwrap_or(false);
-        let autofix = obj
-            .get("autofixable")
-            .and_then(serde_json::Value::as_bool)
-            .unwrap_or(false);
-        let description = obj
-            .get("description")
-            .and_then(serde_json::Value::as_str)
-            .unwrap_or("")
-            .to_owned();
-
-        if enabled {
-            total_enabled += 1;
-        }
-        if autofix {
-            total_autofixable += 1;
-        }
-
-        rows.push([
-            id,
-            name,
-            severity,
-            if enabled {
-                "on".to_owned()
-            } else {
-                "off".to_owned()
-            },
-            if autofix {
-                "yes".to_owned()
-            } else {
-                "no".to_owned()
-            },
-            description,
-        ]);
-    }
-
-    // Column headers.
-    let headers = [
-        "ID",
-        "NAME",
-        "SEVERITY",
-        "ENABLED",
-        "AUTOFIX",
-        "DESCRIPTION",
-    ];
-
-    // Compute max widths for columns 0-4 (description is last, capped).
-    let mut col_widths = [0usize; 6];
-    for (i, h) in headers.iter().enumerate() {
-        col_widths[i] = h.len();
-    }
-    for row in &rows {
-        for i in 0..5 {
-            col_widths[i] = col_widths[i].max(row[i].len());
-        }
-    }
-
-    // Compute remaining width for description column.
-    // Fixed columns + separators (2 spaces between each).
-    let fixed_width: usize = col_widths[..5].iter().sum::<usize>() + 5 * 2; // 5 gaps of 2 spaces
-    let desc_available = term_width.saturating_sub(fixed_width).max(20);
-    let desc_width = desc_available.min(80);
-    col_widths[5] = desc_width;
-
-    let mut s = String::new();
-
-    // Header row.
-    for (i, h) in headers.iter().enumerate() {
-        if i > 0 {
-            s.push_str("  ");
-        }
-        if i == 5 {
-            s.push_str(h);
-        } else {
-            let _ = write!(s, "{h:<width$}", width = col_widths[i]);
-        }
-    }
-    s.push('\n');
-
-    // Data rows.
-    for row in &rows {
-        for (i, cell) in row.iter().enumerate() {
-            if i > 0 {
-                s.push_str("  ");
-            }
-            if i == 5 {
-                // Truncate description with ellipsis if needed.
-                if cell.chars().count() > desc_width {
-                    let truncated: String =
-                        cell.chars().take(desc_width.saturating_sub(1)).collect();
-                    let _ = write!(s, "{truncated}…");
-                } else {
-                    s.push_str(cell);
-                }
-            } else {
-                let _ = write!(s, "{cell:<width$}", width = col_widths[i]);
-            }
-        }
-        s.push('\n');
-    }
-
-    // Summary line.
-    let total = arr.len();
-    let total_disabled = total - total_enabled;
-    let _ = write!(
-        s,
-        "\n{total} rules, {total_enabled} enabled, {total_disabled} disabled. {total_autofixable} autofixable."
-    );
-
-    s
-}
-
-/// Map `Some(true)` → `"on"`, `Some(false)` → `"off"`, `None` → `"?"`.
-fn bool_label(b: Option<bool>) -> &'static str {
-    match b {
-        Some(true) => "on",
-        Some(false) => "off",
-        None => "?",
-    }
-}
-
-/// Format a `lint-rules set` or `lint-rules remove` result as human-readable text.
-///
-/// Expected JSON shapes:
-/// - set:    `{action: "set",    rule_id, dry_run, before: {enabled, severity}, after: {enabled, severity}, config_path}`
-/// - remove: `{action: "remove", rule_id, dry_run, removed, before: {enabled, severity}, after: {enabled, severity}, config_path}`
-///
-/// Rendered example:
-/// ```text
-/// HYALO002: enabled on → off (warn → warn)
-///   wrote .hyalo.toml
-/// ```
-fn format_lint_rules_mutation_text(map: &serde_json::Map<String, serde_json::Value>) -> String {
-    use std::fmt::Write as _;
-
-    let action = map
-        .get("action")
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or("set");
-    let rule_id = map
-        .get("rule_id")
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or("?");
-    let dry_run = map
-        .get("dry_run")
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or(false);
-    let config_path = map
-        .get("config_path")
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or(".hyalo.toml");
-
-    let before_enabled = map
-        .get("before")
-        .and_then(|b| b.get("enabled"))
-        .and_then(serde_json::Value::as_bool);
-    let before_severity = map
-        .get("before")
-        .and_then(|b| b.get("severity"))
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or("warn");
-    let after_enabled = map
-        .get("after")
-        .and_then(|b| b.get("enabled"))
-        .and_then(serde_json::Value::as_bool);
-    let after_severity = map
-        .get("after")
-        .and_then(|b| b.get("severity"))
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or("warn");
-
-    let mut s = String::new();
-
-    if action == "remove" {
-        let removed = map
-            .get("removed")
-            .and_then(serde_json::Value::as_bool)
-            .unwrap_or(false);
-        if !removed {
-            let reason = map
-                .get("reason")
-                .and_then(serde_json::Value::as_str)
-                .unwrap_or("no override found");
-            let _ = write!(s, "{rule_id}: no override to remove ({reason})");
-            return s;
-        }
-        // Show what changed when removing (reverts to defaults).
-        let en_changed = before_enabled != after_enabled;
-        let sev_changed = before_severity != after_severity;
-        let _ = write!(s, "{rule_id}: removed override");
-        if en_changed || sev_changed {
-            let _ = write!(s, " ({}", bool_label(before_enabled));
-            if sev_changed {
-                let _ = write!(s, ", {before_severity}");
-            }
-            let _ = write!(s, " → {}", bool_label(after_enabled));
-            if sev_changed {
-                let _ = write!(s, ", {after_severity}");
-            }
-            s.push(')');
-        }
-    } else {
-        // set action: show enabled and severity changes side-by-side.
-        let en_changed = before_enabled != after_enabled;
-        let sev_changed = before_severity != after_severity;
-
-        let _ = write!(s, "{rule_id}:");
-        if en_changed {
-            let _ = write!(
-                s,
-                " enabled {} → {}",
-                bool_label(before_enabled),
-                bool_label(after_enabled)
-            );
-        }
-        if sev_changed {
-            if en_changed {
-                s.push(',');
-            }
-            let _ = write!(s, " {before_severity} → {after_severity}");
-        }
-        if !en_changed && !sev_changed {
-            s.push_str(" (no change)");
-        }
-    }
-
-    // Indicate write vs dry-run vs no-op.
-    let wrote = map
-        .get("wrote")
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or(true);
-    s.push('\n');
-    if dry_run {
-        if wrote {
-            let _ = write!(s, "  dry-run, would write {config_path}");
-        } else {
-            let _ = write!(s, "  dry-run, no changes to {config_path}");
-        }
-    } else if wrote {
-        let _ = write!(s, "  wrote {config_path}");
-    } else {
-        let _ = write!(s, "  {config_path} unchanged");
-    }
-
-    s
-}
-
-/// Format a `types show` result as human-readable text.
-///
-/// Expected JSON shape: `{type, required, filename_template, defaults, properties}`.
-/// Output example:
-/// ```text
-/// Type: iteration
-///
-/// Required: title, type, date
-///
-/// Properties:
-///   branch:
-///     type: string
-///     pattern: ^iter-\d+/
-///
-///   date:
-///     type: date
-///
-/// Filename template: iteration-{N}-{slug}.md
-/// ```
-fn format_type_show_text(map: &serde_json::Map<String, serde_json::Value>) -> String {
-    use std::fmt::Write as _;
-
-    let mut s = String::new();
-
-    let type_name = map
-        .get("type")
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or("?");
-    let _ = write!(s, "Type: {type_name}");
-
-    // Required fields.
-    if let Some(serde_json::Value::Array(req)) = map.get("required")
-        && !req.is_empty()
-    {
-        let list: Vec<&str> = req.iter().filter_map(serde_json::Value::as_str).collect();
-        let _ = write!(s, "\n\nRequired: {}", list.join(", "));
-    }
-
-    // Defaults block.
-    if let Some(serde_json::Value::Object(defaults)) = map.get("defaults")
-        && !defaults.is_empty()
-    {
-        let _ = write!(s, "\n\nDefaults:");
-        let mut keys: Vec<&str> = defaults.keys().map(String::as_str).collect();
-        keys.sort_unstable();
-        for key in keys {
-            if let Some(value) = defaults.get(key) {
-                let display = match value {
-                    serde_json::Value::String(sv) => sv.clone(),
-                    other => other.to_string(),
-                };
-                let _ = write!(s, "\n  {key}: {display}");
-            }
-        }
-    }
-
-    // Properties block.
-    if let Some(serde_json::Value::Object(props)) = map.get("properties")
-        && !props.is_empty()
-    {
-        let _ = write!(s, "\n\nProperties:");
-        let mut prop_names: Vec<&str> = props.keys().map(String::as_str).collect();
-        prop_names.sort_unstable();
-        for name in prop_names {
-            let Some(prop_val) = props.get(name) else {
-                continue;
-            };
-            let _ = write!(s, "\n  {name}:");
-            if let Some(obj) = prop_val.as_object() {
-                // Print each constraint key on its own indented line.
-                // Always show "type" first, then remaining keys sorted.
-                let mut keys: Vec<&str> = obj.keys().map(String::as_str).collect();
-                keys.sort_unstable_by(|a, b| {
-                    if *a == "type" {
-                        std::cmp::Ordering::Less
-                    } else if *b == "type" {
-                        std::cmp::Ordering::Greater
-                    } else {
-                        a.cmp(b)
-                    }
-                });
-                for key in keys {
-                    if let Some(v) = obj.get(key) {
-                        let display = match v {
-                            serde_json::Value::Array(arr) => arr
-                                .iter()
-                                .filter_map(serde_json::Value::as_str)
-                                .collect::<Vec<_>>()
-                                .join(", "),
-                            serde_json::Value::String(sv) => sv.clone(),
-                            other => other.to_string(),
-                        };
-                        let _ = write!(s, "\n    {key}: {display}");
-                    }
-                }
-            }
-            s.push('\n'); // blank line between property blocks
-        }
-    }
-
-    // Required sections.
-    if let Some(serde_json::Value::Array(sections)) = map.get("required_sections")
-        && !sections.is_empty()
-    {
-        let _ = write!(s, "\n\nRequired sections:");
-        for sec in sections {
-            if let Some(sv) = sec.as_str() {
-                let _ = write!(s, "\n  {sv}");
-            }
-        }
-    }
-
-    // Optional filename template.
-    if let Some(serde_json::Value::String(tmpl)) = map.get("filename_template") {
-        let _ = write!(s, "\nFilename template: {tmpl}");
-    }
-
-    s
-}
-
-/// Format a single `types list` entry as human-readable text.
-///
-/// Expected JSON shape: `{type, required, property_count, has_filename_template}`.
-/// Output example:
-/// ```text
-/// iteration (4 required, 6 properties)
-///   required: title, type, date, tags
-/// ```
-///
-/// Note: `has_filename_template` is a boolean; the actual template is only in `types show`.
-/// When present, a hint to run `types show` is appended.
-fn format_type_list_entry_text(map: &serde_json::Map<String, serde_json::Value>) -> String {
-    use std::fmt::Write as _;
-
-    let mut s = String::new();
-
-    let type_name = map
-        .get("type")
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or("?");
-
-    let req_arr: &[serde_json::Value] = map
-        .get("required")
-        .and_then(serde_json::Value::as_array)
-        .map_or(&[], Vec::as_slice);
-    let req_count = req_arr.len();
-
-    let prop_count = map
-        .get("property_count")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-
-    let has_filename = map
-        .get("has_filename_template")
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or(false);
-
-    let prop_label = if prop_count == 1 {
-        "property"
-    } else {
-        "properties"
-    };
-    let _ = write!(
-        s,
-        "{type_name} ({prop_count} {prop_label}, {req_count} required)"
-    );
-
-    if !req_arr.is_empty() {
-        let list: Vec<&str> = req_arr
-            .iter()
-            .filter_map(serde_json::Value::as_str)
-            .collect();
-        let _ = write!(s, "\n  required: {}", list.join(", "));
-    }
-
-    if has_filename {
-        let _ = write!(s, "\n  filename: (see type details)");
-    }
-
-    s
-}
-
-/// Format a generator command result (`okf index`, `okf log`, `madr toc`) as
-/// readable text. Returns `None` for an unrecognized command so the caller
-/// falls back to the generic renderer.
-fn format_generator_output_text(
-    command: &str,
-    map: &serde_json::Map<String, serde_json::Value>,
-) -> Option<String> {
-    match command {
-        "okf index" => Some(format_okf_index_text(map)),
-        "okf log" => Some(format_okf_log_text(map)),
-        "madr toc" => Some(format_madr_toc_text(map)),
-        _ => None,
-    }
-}
-
-/// Render an `okf index` result: a header line plus one line per changed file
-/// (`  <action> <path>[ (preserving N existing lines)]`).
-fn format_okf_index_text(map: &serde_json::Map<String, serde_json::Value>) -> String {
-    let apply = map.get("apply").and_then(serde_json::Value::as_bool) == Some(true);
-    let changed = map
-        .get("changed")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    let scanned = map
-        .get("scanned")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    let skipped = map
-        .get("skipped_malformed")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-
-    let skipped_markers = map
-        .get("skipped_markers")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    let write_failures = map
-        .get("write_failures")
-        .and_then(serde_json::Value::as_array)
-        .map_or(0, Vec::len) as u64;
-
-    // "written" reads correctly for both singular and plural ("1 file written",
-    // "3 files written"), unlike "1 file wrote".
-    let verb = if apply { "written" } else { "would change" };
-    let mut out = String::new();
-    let _ = write!(
-        out,
-        "okf index: {changed} {} {verb} ({scanned} scanned)",
-        if changed == 1 { "file" } else { "files" }
-    );
-    if skipped > 0 {
-        let _ = write!(out, ", {skipped} skipped (malformed frontmatter)");
-    }
-    if skipped_markers > 0 {
-        let _ = write!(out, ", {skipped_markers} skipped (malformed markers)");
-    }
-    if write_failures > 0 {
-        let _ = write!(out, ", {write_failures} write failure(s)");
-    }
-    append_generator_file_lines(&mut out, map.get("files"));
-    out
-}
-
-/// Append one indented `  <action> <path>` line per entry in a generator's
-/// `files` array (used by `okf index`).
-fn append_generator_file_lines(out: &mut String, files: Option<&serde_json::Value>) {
-    let Some(arr) = files.and_then(serde_json::Value::as_array) else {
-        return;
-    };
-    for f in arr {
-        let Some(obj) = f.as_object() else { continue };
-        let file = obj
-            .get("file")
-            .and_then(serde_json::Value::as_str)
-            .unwrap_or("?");
-        let action = obj
-            .get("action")
-            .and_then(serde_json::Value::as_str)
-            .unwrap_or("change");
-        let _ = write!(out, "\n  {action} {file}");
-        if let Some(n) = obj
-            .get("preserved_lines")
-            .and_then(serde_json::Value::as_u64)
-        {
-            let _ = write!(
-                out,
-                " (preserving {n} existing {})",
-                if n == 1 { "line" } else { "lines" }
-            );
-        }
-        if let Some(reason) = obj.get("reason").and_then(serde_json::Value::as_str) {
-            let _ = write!(out, " — {reason}");
-        }
-    }
-}
-
-/// Render an `okf log` result: a single line describing the written entry.
-fn format_okf_log_text(map: &serde_json::Map<String, serde_json::Value>) -> String {
-    let apply = map.get("apply").and_then(serde_json::Value::as_bool) == Some(true);
-    let file = map
-        .get("file")
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or("log.md");
-    let date = map
-        .get("date")
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or("");
-    let created = map.get("created").and_then(serde_json::Value::as_bool) == Some(true);
-    let verb = if apply { "logged" } else { "would log" };
-    let mut out = format!("okf log: {verb} entry under {date} in {file}");
-    if created {
-        out.push_str(" (created)");
-    }
-    if let Some(entry) = map.get("entry").and_then(serde_json::Value::as_str) {
-        let _ = write!(out, "\n  {entry}");
-    }
-    out
-}
-
-/// Render a `madr toc` result: a single line describing the TOC action.
-fn format_madr_toc_text(map: &serde_json::Map<String, serde_json::Value>) -> String {
-    let apply = map.get("apply").and_then(serde_json::Value::as_bool) == Some(true);
-    let file = map
-        .get("file")
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or("README.md");
-    let action = map
-        .get("action")
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or("unchanged");
-    let adrs = map
-        .get("adrs")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-
-    let mut out = if apply {
-        format!("madr toc: {action} {file} ({adrs} ADRs)")
-    } else if action == "unchanged" {
-        format!("madr toc: {file} up to date ({adrs} ADRs)")
-    } else {
-        format!("madr toc: would {action} {file} ({adrs} ADRs)")
-    };
-    if let Some(n) = map
-        .get("preserved_lines")
-        .and_then(serde_json::Value::as_u64)
-    {
-        let _ = write!(
-            out,
-            " (preserving {n} existing {})",
-            if n == 1 { "line" } else { "lines" }
-        );
-    }
-    out
-}
-
-/// Generic key: value rendering for unknown object shapes.
-fn format_object_generic(
-    map: &serde_json::Map<String, serde_json::Value>,
-    cache: &mut JaqFilterCache,
-) -> String {
-    map.iter()
-        .map(|(k, v)| format!("{k}: {}", format_value_as_text(v, cache)))
-        .collect::<Vec<_>>()
-        .join("\n")
-}
-
-/// Format a scalar JSON value as text.
-fn format_scalar(value: &serde_json::Value, cache: &mut JaqFilterCache) -> String {
-    match value {
-        serde_json::Value::String(s) => s.clone(),
-        serde_json::Value::Number(n) => n.to_string(),
-        serde_json::Value::Bool(b) => b.to_string(),
-        serde_json::Value::Null => "null".to_owned(),
-        serde_json::Value::Array(arr) => {
-            let items: Vec<String> = arr.iter().map(|v| format_scalar(v, cache)).collect();
-            items.join(", ")
-        }
-        serde_json::Value::Object(_) => format_value_as_text(value, cache),
-    }
-}
+//
+// `output.rs` had grown to 3,744 lines -- a review hotspot flagged by the
+// 2026-08-27 deep review. It is split along the pipeline: the static filter
+// table, the jq engine, the dynamic file-object filter, and the hand-written
+// text renderings. This file keeps the envelope and the `Format` /
+// `CommandOutcome` vocabulary every command speaks.
+
+mod file_object;
+mod filters;
+mod jq;
+mod text;
+mod text_generators;
+mod text_lint;
+mod text_types;
+
+// Glob re-exports: each item keeps the visibility it had before the split (a
+// glob re-export cannot widen it), and this is also how the submodules reach
+// each other, since each one imports this parent.
+use file_object::build_file_object_filter;
+use filters::{key_signature, lookup_filter};
+pub use jq::*;
+use text::format_value_as_text;
+use text_generators::format_generator_output_text;
+use text_lint::{
+    format_lint_fix_output_text, format_lint_output_text, format_lint_rules_list_text,
+    format_lint_rules_mutation_text,
+};
+use text_types::{format_type_list_entry_text, format_type_show_text};
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use serde_json::json;
-
-    // Convenience wrappers so individual tests don't have to construct a cache.
-    fn jq(filter: &str, val: &serde_json::Value) -> Option<String> {
-        apply_jq_filter(filter, val, &mut JaqFilterCache::new())
-    }
-
-    fn fmt(val: &serde_json::Value) -> String {
-        format_value_as_text(val, &mut JaqFilterCache::new())
-    }
-
-    fn scalar(val: &serde_json::Value) -> String {
-        format_scalar(val, &mut JaqFilterCache::new())
-    }
-
-    // -----------------------------------------------------------------------
-    // iter-213 UX-5 — a rule is never shown as both fixed and conflicted
-    // -----------------------------------------------------------------------
-
-    fn fix_output(fixed_rule: Option<&str>, conflict_rule: &str) -> String {
-        let fixed_groups = match fixed_rule {
-            Some(rule) => json!([{"rule": rule, "count": 1, "violations": []}]),
-            None => json!([]),
-        };
-        let value = json!({
-            "dry_run": false,
-            "files_checked": 1,
-            "total_fixed": 1,
-            "total_remaining": 0,
-            "total_conflicts": 1,
-            "files": [{
-                "file": "a.md",
-                "fixed_groups": fixed_groups,
-                "remaining_groups": [],
-                "conflicts": [{"rule": conflict_rule, "reason": "range overlap with MD009"}],
-            }],
-        });
-        format_lint_fix_output_text(value.as_object().expect("object"))
-    }
-
-    #[test]
-    fn fix_text_suppresses_a_conflict_for_a_rule_already_shown_as_fixed() {
-        let out = fix_output(Some("MD047"), "MD047");
-        assert!(
-            !out.contains("conflict  MD047"),
-            "a rule shown as fixed must not also be shown as conflicted: {out}"
-        );
-        assert!(
-            out.contains("MD047"),
-            "the fixed line itself must survive: {out}"
-        );
-    }
-
-    #[test]
-    fn fix_text_keeps_a_conflict_for_a_rule_that_was_not_fixed() {
-        let out = fix_output(Some("MD047"), "MD013");
-        assert!(
-            out.contains("conflict  MD013"),
-            "an unrelated rule's conflict is the whole point of the line: {out}"
-        );
-    }
-
-    // --- error formatting ---
-
-    #[test]
-    fn format_json_error() {
-        let out = format_error(
-            Format::Json,
-            "file not found",
-            Some("foo/bar"),
-            Some("did you mean foo/bar.md?"),
-            None,
-        );
-        let parsed: serde_json::Value = serde_json::from_str(&out).unwrap();
-        assert_eq!(parsed["error"], "file not found");
-        assert_eq!(parsed["hint"], "did you mean foo/bar.md?");
-        assert!(parsed.get("cause").is_none());
-    }
-
-    #[test]
-    fn format_text_error() {
-        let out = format_error(Format::Text, "file not found", Some("foo"), None, None);
-        assert!(out.contains("Error: file not found"));
-        assert!(out.contains("path: foo"));
-    }
-
-    #[test]
-    fn format_json_success() {
-        let val = json!({"name": "test", "value": 42});
-        let out = format_success(Format::Json, &val);
-        assert!(out.contains("\"name\": \"test\""));
-    }
-
-    // --- apply_jq_filter ---
-
-    #[test]
-    fn apply_jq_filter_simple() {
-        let val = json!({"name": "hello", "count": 3});
-        let result = jq(r#""\(.name): \(.count)""#, &val);
-        assert_eq!(result.as_deref(), Some("hello: 3"));
-    }
-
-    #[test]
-    fn apply_jq_filter_array_map() {
-        let val = json!(["a", "b", "c"]);
-        let result = jq(".[]", &val);
-        assert_eq!(result.as_deref(), Some("a\nb\nc"));
-    }
-
-    #[test]
-    fn apply_jq_filter_invalid_returns_none() {
-        let val = json!({"x": 1});
-        let result = jq("this is not valid jq %%%", &val);
-        assert!(result.is_none());
-    }
-
-    // --- F-3: jq runtime errors are bounded, not whole-input dumps ---
-
-    #[test]
-    fn jq_runtime_error_on_large_input_is_bounded_and_names_the_filter() {
-        // A filter that fails against a large array (e.g. mistakenly indexing
-        // an array with a string, the `--jq '.results | .file'` shape from
-        // deep-analysis-2's F-3) must not embed the whole array — jaq's
-        // runtime Display would otherwise dump every element into the error.
-        let big_array: Vec<serde_json::Value> = (0..5000)
-            .map(|i| json!({"file": format!("file-{i}.md"), "padding": "x".repeat(200)}))
-            .collect();
-        let val = json!(big_array);
-        let err = apply_jq_filter_result(".file", &val).expect_err("indexing an array must fail");
-
-        assert!(
-            err.len() < 1000,
-            "error should be bounded to roughly 2x the char cap, got {} bytes",
-            err.len()
-        );
-        assert!(
-            err.contains(".file"),
-            "error should name the failing filter: {err}"
-        );
-        assert!(
-            !err.contains("file-4999.md"),
-            "error must not embed the tail of the huge input value: {err}"
-        );
-    }
-
-    #[test]
-    fn jq_runtime_error_on_large_object_value_does_not_leak_its_content() {
-        // Array-indexing an object fails at runtime; the object (which holds
-        // a large string field) must not appear verbatim in the error.
-        let mut map = serde_json::Map::new();
-        let needle = "x".repeat(5000);
-        map.insert("huge".to_owned(), serde_json::Value::String(needle.clone()));
-        let val = serde_json::Value::Object(map);
-
-        let err = apply_jq_filter_result(".huge.nope", &val)
-            .expect_err("field-indexing a string value must fail");
-        assert!(
-            err.len() < 1000,
-            "error should be bounded, got {} bytes",
-            err.len()
-        );
-        assert!(
-            !err.contains(&needle),
-            "error must not embed the large field value verbatim"
-        );
-    }
-
-    #[test]
-    fn truncate_diagnostic_appends_ellipsis_only_when_truncated() {
-        let short = "short message";
-        assert_eq!(truncate_diagnostic(short), short);
-
-        let long = "x".repeat(500);
-        let truncated = truncate_diagnostic(&long);
-        assert_eq!(truncated.chars().count(), JQ_ERROR_DIAGNOSTIC_CHAR_CAP + 1);
-        assert!(truncated.ends_with('…'));
-    }
-
-    #[test]
-    fn truncate_diagnostic_never_splits_a_multibyte_codepoint() {
-        // 300 multi-byte CJK characters — truncation must cut on char
-        // boundaries, not bytes, or this would panic on a split codepoint.
-        let long: String = "日".repeat(300);
-        let truncated = truncate_diagnostic(&long);
-        assert_eq!(truncated.chars().count(), JQ_ERROR_DIAGNOSTIC_CHAR_CAP + 1);
-        assert!(truncated.ends_with('…'));
-    }
-
-    // --- jq output size cap ---
-
-    #[test]
-    fn jq_output_cap_constant_is_10_mib() {
-        assert_eq!(JQ_OUTPUT_CAP, 10 * 1024 * 1024);
-    }
-
-    #[test]
-    fn jq_output_within_cap_succeeds() {
-        // A small output must pass through without hitting the cap.
-        let val = json!({"msg": "hello"});
-        let result = apply_jq_filter_result(".msg", &val);
-        assert_eq!(result.as_deref(), Ok("hello"));
-    }
-
-    #[test]
-    fn jq_output_cap_triggers_on_large_output() {
-        // Build a JSON array large enough to exceed JQ_OUTPUT_CAP when expanded.
-        // Each element is "aaaa...a" (1000 chars). 11_000 elements = 11 MB > 10 MB cap.
-        let big_string = "a".repeat(1000);
-        let val = serde_json::Value::Array(
-            std::iter::repeat_n(serde_json::Value::String(big_string), 11_000).collect(),
-        );
-        // ".[]" emits each element as a separate output value.
-        let result = apply_jq_filter_result(".[]", &val);
-        assert!(result.is_err(), "expected cap error but got Ok output");
-        let err = result.unwrap_err();
-        assert!(
-            err.contains("exceeds") && err.contains("MiB"),
-            "unexpected error message: {err}"
-        );
-    }
-
-    #[test]
-    fn jq_single_huge_string_value_rejected_via_raw_length_before_copying() {
-        // Finding 2 (review round on PR #254): a single value already over
-        // JQ_OUTPUT_CAP on its own must be rejected using its *existing*
-        // borrowed bytes, not measured only after being duplicated into an
-        // owned copy. Uses an 11 MiB *input* string (cheap to build in a
-        // unit test) piped through the identity filter `.` — this emits it
-        // as a single `Val::TStr`, exercising the exact single-value
-        // pre-check path the real repro (`"x" * 2000000000`, ~4 GB
-        // unmitigated) hits, without needing gigabytes in a fast test.
-        let big_string = "a".repeat(11 * 1024 * 1024);
-        let val = serde_json::Value::String(big_string);
-        let result = apply_jq_filter_result(".", &val);
-        assert!(
-            result.is_err(),
-            "a single value over the byte cap must be rejected"
-        );
-        let err = result.unwrap_err();
-        assert!(
-            err.contains("exceeds") && err.contains("MiB"),
-            "unexpected error message: {err}"
-        );
-    }
-
-    // --- F3-1: jq resource limits (deep-analysis-3-2026-08-23.md) ---
-
-    #[test]
-    fn jq_max_output_values_constant_is_one_million() {
-        assert_eq!(JQ_MAX_OUTPUT_VALUES, 1_000_000);
-    }
-
-    #[test]
-    fn jq_time_limit_constant_is_three_seconds() {
-        assert_eq!(JQ_TIME_LIMIT, std::time::Duration::from_secs(3));
-    }
-
-    #[test]
-    fn jq_value_count_cap_triggers_before_byte_cap_on_many_tiny_values() {
-        // `range(...)` without array-collection is a streaming generator: our
-        // for loop in `execute_jq_filter` pulls one value at a time, so this
-        // is cheap (no huge intermediate) and exercises JQ_MAX_OUTPUT_VALUES
-        // directly rather than JQ_TIME_LIMIT or JQ_OUTPUT_CAP. Each emitted
-        // value is a single-digit-or-more number (a few bytes), so the byte
-        // cap (10 MiB) would take far longer than 1,000,000 iterations to
-        // reach — the value-count cap must fire first.
-        let val = json!(null);
-        let result = apply_jq_filter_result("range(2000000)", &val);
-        assert!(
-            result.is_err(),
-            "expected the value-count cap to trigger, got Ok"
-        );
-        let err = result.unwrap_err();
-        assert!(
-            err.contains("1000000") && err.contains("values"),
-            "expected a value-count-limit error, got: {err}"
-        );
-    }
-
-    #[test]
-    fn jq_infinite_recursion_with_no_output_errors_within_time_limit() {
-        // `def f: f; f` never emits a value at all, so the loop body in
-        // `execute_jq_filter` never runs even once — neither JQ_OUTPUT_CAP
-        // nor JQ_MAX_OUTPUT_VALUES can catch it, only the wall-clock
-        // deadline on the worker thread can. This genuinely blocks for
-        // ~JQ_TIME_LIMIT (3s) — an acceptable one-time cost for a real
-        // regression test on a HIGH-severity finding; CPU cost is trivial
-        // (no allocation), unlike the array-collection case covered by the
-        // e2e tests in `tests/e2e/jq.rs`.
-        let val = json!(null);
-        let started = std::time::Instant::now();
-        let result = apply_jq_filter_result("def f: f; f", &val);
-        assert!(
-            result.is_err(),
-            "expected a time-limit error, got Ok output"
-        );
-        let err = result.unwrap_err();
-        assert!(
-            err.contains("time limit"),
-            "expected a time-limit error, got: {err}"
-        );
-        assert!(
-            started.elapsed() < std::time::Duration::from_secs(9),
-            "must return well within a small multiple of JQ_TIME_LIMIT, not hang"
-        );
-    }
-
-    // --- property type filters ---
-
-    #[test]
-    fn property_info_filter() {
-        let val = json!({"name": "title", "type": "text", "value": "My Note"});
-        let out = jq(PROPERTY_INFO_FILTER, &val).unwrap();
-        assert!(out.contains("title"));
-        assert!(out.contains("text"));
-        assert!(out.contains("My Note"));
-    }
-
-    #[test]
-    fn property_info_filter_list_value() {
-        let val = json!({"name": "tags", "type": "list", "value": ["rust", "cli"]});
-        let out = jq(PROPERTY_INFO_FILTER, &val).unwrap();
-        assert!(out.contains("tags"));
-        assert!(out.contains("list"));
-        // Array values should be wrapped in brackets and joined with ", "
-        assert!(out.contains("[rust, cli]"), "expected [rust, cli]: {out}");
-        assert!(!out.contains("[\"rust\""));
-    }
-
-    #[test]
-    fn property_summary_entry_filter() {
-        let val = json!({"count": 7, "name": "title", "type": "text"});
-        let out = jq(PROPERTY_SUMMARY_ENTRY_FILTER, &val).unwrap();
-        assert!(out.contains("title"));
-        assert!(out.contains("text"));
-        assert!(out.contains("7 files"));
-    }
-
-    #[test]
-    fn tag_summary_filter() {
-        let val = json!({
-            "tags": [{"name": "rust", "count": 3}, {"name": "cli", "count": 1}],
-            "total": 2
-        });
-        let out = jq(TAG_SUMMARY_FILTER, &val).unwrap();
-        assert!(out.contains("2 unique tags"));
-        assert!(out.contains("rust"));
-        assert!(out.contains("3 files"));
-    }
-
-    // --- link type filters ---
-
-    #[test]
-    fn link_info_target_only_filter() {
-        let val = json!({"target": "broken-link"});
-        let out = jq(LINK_INFO_TARGET_FILTER, &val).unwrap();
-        assert!(out.contains("broken-link"));
-        assert!(out.contains("unresolved"));
-    }
-
-    #[test]
-    fn link_info_with_path_filter() {
-        let val = json!({"path": "note-b.md", "target": "note-b"});
-        let out = jq(LINK_INFO_PATH_FILTER, &val).unwrap();
-        assert!(out.contains("note-b"));
-        assert!(out.contains("note-b.md"));
-    }
-
-    #[test]
-    fn link_info_anchored_resolved_filter() {
-        // Target and anchor both resolve.
-        let val = json!({"path": "foo.md", "target": "Foo", "fragment": "Real"});
-        let out = jq(LINK_INFO_ANCHORED_FILTER, &val).unwrap();
-        assert!(out.contains("Foo#Real"));
-        assert!(out.contains("foo.md"));
-        assert!(!out.contains("broken anchor"));
-        assert!(!out.contains("unresolved"));
-    }
-
-    #[test]
-    fn link_info_anchored_broken_anchor_filter() {
-        // Target resolves but the anchor does not — distinct from unresolved.
-        let val = json!({
-            "path": "foo.md", "target": "Foo", "fragment": "Nope", "broken_anchor": true
-        });
-        let out = jq(LINK_INFO_ANCHORED_FILTER, &val).unwrap();
-        assert!(out.contains("Foo#Nope"));
-        assert!(out.contains("foo.md"));
-        assert!(out.contains("broken anchor"));
-        assert!(!out.contains("unresolved"));
-    }
-
-    #[test]
-    fn link_info_anchored_broken_target_filter() {
-        // Target unresolved: anchor check never ran, path is null.
-        let val = json!({"target": "Nope", "fragment": "x"});
-        let out = jq(LINK_INFO_ANCHORED_FILTER, &val).unwrap();
-        assert!(out.contains("Nope#x"));
-        assert!(out.contains("unresolved"));
-        assert!(!out.contains("broken anchor"));
-    }
-
-    #[test]
-    fn link_info_anchored_with_label_filter() {
-        let val = json!({
-            "path": "foo.md", "target": "Foo", "fragment": "Real", "label": "click"
-        });
-        let out = jq(LINK_INFO_ANCHORED_FILTER, &val).unwrap();
-        assert!(out.contains("Foo#Real"));
-        assert!(out.contains("[click]"));
-    }
-
-    // --- outline type filters ---
-
-    #[test]
-    fn task_count_filter() {
-        let val = json!({"done": 3, "total": 5});
-        let out = jq(TASK_COUNT_FILTER, &val).unwrap();
-        assert_eq!(out, "[3/5]");
-    }
-
-    #[test]
-    fn outline_section_filter() {
-        let val = json!({
-            "code_blocks": [],
-            "heading": "Introduction",
-            "level": 1,
-            "line": 5,
-            "links": ["[[other]]"]
-        });
-        let out = jq(OUTLINE_SECTION_FILTER, &val).unwrap();
-        assert!(out.contains('#'));
-        assert!(out.contains("Introduction"));
-        assert!(out.contains("[[other]]"));
-    }
-
-    #[test]
-    fn outline_section_with_tasks_filter() {
-        let val = json!({
-            "code_blocks": [],
-            "heading": "Tasks",
-            "level": 2,
-            "line": 10,
-            "links": [],
-            "tasks": {"done": 2, "total": 4}
-        });
-        let out = jq(OUTLINE_SECTION_WITH_TASKS_FILTER, &val).unwrap();
-        assert!(out.contains("##"));
-        assert!(out.contains("Tasks"));
-        assert!(out.contains("[2/4]"));
-    }
-
-    /// NEW-16 (dogfood pre3): a hand-written `[n/m]` already in the heading
-    /// text must render once — the computed count replaces it rather than
-    /// appending a second bracket group (`## Tasks [6/6] [2/4]`).
-    #[test]
-    fn outline_section_with_tasks_filter_replaces_a_stale_hand_written_count() {
-        let val = json!({
-            "code_blocks": [],
-            "heading": "Tasks [6/6]",
-            "level": 2,
-            "line": 10,
-            "links": [],
-            "tasks": {"done": 2, "total": 4}
-        });
-        let out = jq(OUTLINE_SECTION_WITH_TASKS_FILTER, &val).unwrap();
-        assert!(
-            out.contains("Tasks [2/4]"),
-            "expected the computed count, got: {out}"
-        );
-        assert!(
-            !out.contains("[6/6]"),
-            "the stale hand-written count must not survive: {out}"
-        );
-        assert_eq!(
-            out.matches('[').count(),
-            1,
-            "exactly one bracket group must render: {out}"
-        );
-    }
-
-    // --- FindTaskInfo filter ---
-
-    #[test]
-    fn find_task_info_filter_done() {
-        let val = json!({
-            "done": true,
-            "line": 42,
-            "section": "Implementation",
-            "status": "x",
-            "text": "Write the tests"
-        });
-        let out = jq(FIND_TASK_INFO_FILTER, &val).unwrap();
-        assert!(out.contains("[x]"));
-        assert!(out.contains("Write the tests"));
-        assert!(out.contains("line 42"));
-        assert!(out.contains("Implementation"));
-    }
-
-    #[test]
-    fn find_task_info_filter_not_done() {
-        let val = json!({
-            "done": false,
-            "line": 7,
-            "section": "Todo",
-            "status": " ",
-            "text": "Review PR"
-        });
-        let out = jq(FIND_TASK_INFO_FILTER, &val).unwrap();
-        assert!(out.contains("[ ]"));
-        assert!(out.contains("Review PR"));
-        assert!(out.contains("line 7"));
-        assert!(out.contains("Todo"));
-    }
-
-    #[test]
-    fn find_task_info_via_format_value_as_text() {
-        // Verify that format_value_as_text dispatches to the correct filter.
-        let val = json!({
-            "done": true,
-            "line": 5,
-            "section": "Goals",
-            "status": "x",
-            "text": "Ship it"
-        });
-        let out = fmt(&val);
-        assert!(out.contains("[x]"));
-        assert!(out.contains("Ship it"));
-        assert!(
-            !out.contains("done: true"),
-            "should not use generic fallback"
-        );
-    }
-
-    // --- ContentMatch filter ---
-
-    #[test]
-    fn content_match_filter() {
-        let val = json!({
-            "line": 15,
-            "section": "Background",
-            "text": "This is the matching line"
-        });
-        let out = jq(CONTENT_MATCH_FILTER, &val).unwrap();
-        assert!(out.contains("line 15"));
-        assert!(out.contains("Background"));
-        assert!(out.contains("This is the matching line"));
-    }
-
-    #[test]
-    fn content_match_via_format_value_as_text() {
-        let val = json!({
-            "line": 3,
-            "section": "Intro",
-            "text": "hello world"
-        });
-        let out = fmt(&val);
-        assert!(out.contains("line 3"));
-        assert!(out.contains("hello world"));
-        assert!(!out.contains("line: 3"), "should not use generic fallback");
-    }
-
-    // --- Mutation result filters ---
-
-    #[test]
-    fn property_value_mutation_filter_with_modified() {
-        // SetPropertyResult / AppendPropertyResult / RemovePropertyResult (with value)
-        // scanned == total: no "(N scanned)" suffix
-        let val = json!({
-            "modified": ["note-a.md", "note-b.md"],
-            "property": "status",
-            "scanned": 2,
-            "skipped": [],
-            "skipped_count": 0,
-            "total": 2,
-            "value": "done"
-        });
-        let out = jq(PROPERTY_VALUE_MUTATION_FILTER, &val).unwrap();
-        assert!(out.contains("status=done"));
-        assert!(out.contains("2/2 modified"));
-        assert!(
-            !out.contains("scanned"),
-            "no scanned suffix when scanned == total"
-        );
-        assert!(out.contains("note-a.md"));
-        assert!(out.contains("note-b.md"));
-    }
-
-    #[test]
-    fn property_value_mutation_filter_all_skipped() {
-        let val = json!({
-            "modified": [],
-            "property": "priority",
-            "scanned": 1,
-            "skipped": ["note-a.md"],
-            "skipped_count": 1,
-            "total": 1,
-            "value": "high"
-        });
-        let out = jq(PROPERTY_VALUE_MUTATION_FILTER, &val).unwrap();
-        assert!(out.contains("priority=high"));
-        assert!(out.contains("0/1 modified"));
-        // No file paths should appear when nothing was modified
-        assert!(!out.contains("note-a.md"));
-    }
-
-    #[test]
-    fn property_value_mutation_filter_with_where_filter() {
-        // scanned > total: "(N scanned)" suffix should appear
-        let val = json!({
-            "modified": ["note-a.md"],
-            "property": "status",
-            "scanned": 5,
-            "skipped": [],
-            "skipped_count": 0,
-            "total": 1,
-            "value": "done"
-        });
-        let out = jq(PROPERTY_VALUE_MUTATION_FILTER, &val).unwrap();
-        assert!(out.contains("status=done"));
-        assert!(out.contains("1/1 modified"));
-        assert!(out.contains("(5 scanned)"));
-    }
-
-    #[test]
-    fn property_value_mutation_via_format_value_as_text() {
-        let val = json!({
-            "dry_run": false,
-            "modified": ["notes/a.md"],
-            "property": "status",
-            "scanned": 1,
-            "skipped": [],
-            "skipped_count": 0,
-            "total": 1,
-            "value": "done"
-        });
-        let out = fmt(&val);
-        assert!(out.contains("status=done"));
-        assert!(
-            !out.contains("modified: "),
-            "should not use generic fallback"
-        );
-    }
-
-    #[test]
-    fn property_mutation_filter_no_value() {
-        // RemovePropertyResult without value; scanned == total
-        let val = json!({
-            "dry_run": false,
-            "modified": ["note.md"],
-            "property": "draft",
-            "scanned": 1,
-            "skipped": [],
-            "skipped_count": 0,
-            "total": 1
-        });
-        let out = jq(PROPERTY_MUTATION_FILTER, &val).unwrap();
-        assert!(out.contains("draft"));
-        assert!(out.contains("1/1 modified"));
-        assert!(
-            !out.contains("scanned"),
-            "no scanned suffix when scanned == total"
-        );
-        assert!(out.contains("note.md"));
-    }
-
-    #[test]
-    fn property_mutation_filter_no_value_with_where_filter() {
-        // RemovePropertyResult without value; scanned > total
-        let val = json!({
-            "dry_run": false,
-            "modified": ["note.md"],
-            "property": "draft",
-            "scanned": 7,
-            "skipped": [],
-            "skipped_count": 0,
-            "total": 1
-        });
-        let out = jq(PROPERTY_MUTATION_FILTER, &val).unwrap();
-        assert!(out.contains("draft"));
-        assert!(out.contains("1/1 modified"));
-        assert!(out.contains("(7 scanned)"));
-    }
-
-    #[test]
-    fn tag_mutation_filter_with_modified() {
-        // SetTagResult / RemoveTagResult; scanned == total
-        let val = json!({
-            "dry_run": false,
-            "modified": ["a.md", "b.md"],
-            "scanned": 3,
-            "skipped": ["c.md"],
-            "skipped_count": 1,
-            "tag": "rust",
-            "total": 3
-        });
-        let out = jq(TAG_MUTATION_FILTER, &val).unwrap();
-        assert!(out.contains("rust"));
-        assert!(out.contains("2/3 modified"));
-        assert!(
-            !out.contains("scanned"),
-            "no scanned suffix when scanned == total"
-        );
-        assert!(out.contains("a.md"));
-        assert!(out.contains("b.md"));
-        assert!(!out.contains("c.md"));
-    }
-
-    #[test]
-    fn tag_mutation_filter_with_where_filter() {
-        // scanned > total: "(N scanned)" suffix
-        let val = json!({
-            "dry_run": false,
-            "modified": ["a.md"],
-            "scanned": 10,
-            "skipped": [],
-            "skipped_count": 0,
-            "tag": "rust",
-            "total": 1
-        });
-        let out = jq(TAG_MUTATION_FILTER, &val).unwrap();
-        assert!(out.contains("rust"));
-        assert!(out.contains("1/1 modified"));
-        assert!(out.contains("(10 scanned)"));
-    }
-
-    #[test]
-    fn tag_mutation_via_format_value_as_text() {
-        let val = json!({
-            "dry_run": false,
-            "modified": [],
-            "scanned": 1,
-            "skipped": ["note.md"],
-            "skipped_count": 1,
-            "tag": "cli",
-            "total": 1
-        });
-        let out = fmt(&val);
-        assert!(out.contains("cli"));
-        assert!(!out.contains("tag: cli"), "should not use generic fallback");
-    }
-
-    // --- dry-run prefix in text output ---
-
-    #[test]
-    fn property_value_mutation_dry_run_prefix() {
-        let val = json!({
-            "dry_run": true,
-            "modified": ["note.md"],
-            "property": "status",
-            "scanned": 1,
-            "skipped": [],
-            "skipped_count": 0,
-            "total": 1,
-            "value": "done"
-        });
-        let out = fmt(&val);
-        assert!(
-            out.contains("[dry-run] status=done"),
-            "dry-run prefix missing: {out}"
-        );
-    }
-
-    #[test]
-    fn tag_mutation_dry_run_prefix() {
-        let val = json!({
-            "dry_run": true,
-            "modified": ["note.md"],
-            "scanned": 1,
-            "skipped": [],
-            "skipped_count": 0,
-            "tag": "rust",
-            "total": 1
-        });
-        let out = fmt(&val);
-        assert!(
-            out.contains("[dry-run] rust"),
-            "dry-run prefix missing: {out}"
-        );
-    }
-
-    #[test]
-    fn property_value_mutation_no_dry_run_prefix() {
-        let val = json!({
-            "dry_run": false,
-            "modified": ["note.md"],
-            "property": "status",
-            "scanned": 1,
-            "skipped": [],
-            "skipped_count": 0,
-            "total": 1,
-            "value": "done"
-        });
-        let out = fmt(&val);
-        assert!(
-            !out.contains("[dry-run]"),
-            "should not have dry-run prefix: {out}"
-        );
-    }
-
-    // --- build_file_object_filter ---
-
-    #[test]
-    fn build_file_object_filter_minimal() {
-        // Only the required `file` and `modified` fields.
-        let map: serde_json::Map<String, serde_json::Value> =
-            serde_json::from_str(r#"{"file": "notes/foo.md", "modified": "2024-01-01"}"#).unwrap();
-        let filter = build_file_object_filter(&map);
-        let val = json!({"file": "notes/foo.md", "modified": "2024-01-01"});
-        let out = jq(&filter, &val).unwrap();
-        assert!(out.contains("notes/foo.md"));
-        assert!(out.contains("2024-01-01"));
-    }
-
-    #[test]
-    fn build_file_object_filter_with_tags() {
-        let map: serde_json::Map<String, serde_json::Value> = serde_json::from_str(
-            r#"{"file": "foo.md", "modified": "2024-01-01", "tags": ["rust", "cli"]}"#,
-        )
-        .unwrap();
-        let filter = build_file_object_filter(&map);
-        let val = json!({"file": "foo.md", "modified": "2024-01-01", "tags": ["rust", "cli"]});
-        let out = jq(&filter, &val).unwrap();
-        assert!(out.contains("foo.md"));
-        assert!(out.contains("tags: [rust, cli]"));
-    }
-
-    #[test]
-    fn build_file_object_filter_with_properties() {
-        let map: serde_json::Map<String, serde_json::Value> = serde_json::from_str(
-            r#"{"file": "foo.md", "modified": "2024-01-01", "properties": {"status": "done"}}"#,
-        )
-        .unwrap();
-        let filter = build_file_object_filter(&map);
-        let val = json!({
-            "file": "foo.md",
-            "modified": "2024-01-01",
-            "properties": {"status": "done"}
-        });
-        let out = jq(&filter, &val).unwrap();
-        assert!(out.contains("foo.md"));
-        assert!(out.contains("properties:"));
-        assert!(out.contains("status: done"));
-    }
-
-    #[test]
-    fn build_file_object_filter_with_tasks() {
-        let map: serde_json::Map<String, serde_json::Value> = serde_json::from_str(
-            r#"{"file": "foo.md", "modified": "2024-01-01", "tasks": [{"done": true, "line": 5, "section": "Goals", "status": "x", "text": "Ship it"}]}"#,
-        )
-        .unwrap();
-        let filter = build_file_object_filter(&map);
-        let val = json!({
-            "file": "foo.md",
-            "modified": "2024-01-01",
-            "tasks": [{"done": true, "line": 5, "section": "Goals", "status": "x", "text": "Ship it"}]
-        });
-        let out = jq(&filter, &val).unwrap();
-        assert!(out.contains("foo.md"));
-        assert!(out.contains("tasks:"));
-        assert!(out.contains("[x] Ship it"));
-        assert!(out.contains("line 5"));
-    }
-
-    #[test]
-    fn build_file_object_filter_with_sections() {
-        let map: serde_json::Map<String, serde_json::Value> = serde_json::from_str(
-            r#"{"file": "foo.md", "modified": "2024-01-01", "sections": [{"code_blocks": 0, "heading": "Intro", "level": 1, "line": 1, "links": []}]}"#,
-        )
-        .unwrap();
-        let filter = build_file_object_filter(&map);
-        let val = json!({
-            "file": "foo.md",
-            "modified": "2024-01-01",
-            "sections": [{"code_blocks": 0, "heading": "Intro", "level": 1, "line": 1, "links": []}]
-        });
-        let out = jq(&filter, &val).unwrap();
-        assert!(out.contains("foo.md"));
-        assert!(out.contains("sections:"));
-        assert!(out.contains("# Intro"));
-    }
-
-    #[test]
-    fn build_file_object_filter_with_matches() {
-        let map: serde_json::Map<String, serde_json::Value> = serde_json::from_str(
-            r#"{"file": "foo.md", "modified": "2024-01-01", "matches": [{"line": 3, "section": "Intro", "text": "hello world"}]}"#,
-        )
-        .unwrap();
-        let filter = build_file_object_filter(&map);
-        let val = json!({
-            "file": "foo.md",
-            "modified": "2024-01-01",
-            "matches": [{"line": 3, "section": "Intro", "text": "hello world"}]
-        });
-        let out = jq(&filter, &val).unwrap();
-        assert!(out.contains("foo.md"));
-        assert!(out.contains("matches:"));
-        assert!(out.contains("line 3 (Intro): hello world"));
-    }
-
-    #[test]
-    fn build_file_object_filter_with_links() {
-        let map: serde_json::Map<String, serde_json::Value> = serde_json::from_str(
-            r#"{"file": "foo.md", "modified": "2024-01-01", "links": [{"target": "bar", "path": "bar.md"}]}"#,
-        )
-        .unwrap();
-        let filter = build_file_object_filter(&map);
-        let val = json!({
-            "file": "foo.md",
-            "modified": "2024-01-01",
-            "links": [{"target": "bar", "path": "bar.md"}]
-        });
-        let out = jq(&filter, &val).unwrap();
-        assert!(out.contains("foo.md"));
-        assert!(out.contains("links:"));
-        assert!(out.contains(r#""bar" → "bar.md""#));
-    }
-
-    #[test]
-    fn build_file_object_filter_unresolved_link() {
-        let map: serde_json::Map<String, serde_json::Value> = serde_json::from_str(
-            r#"{"file": "foo.md", "modified": "2024-01-01", "links": [{"target": "missing"}]}"#,
-        )
-        .unwrap();
-        let filter = build_file_object_filter(&map);
-        let val = json!({
-            "file": "foo.md",
-            "modified": "2024-01-01",
-            "links": [{"target": "missing"}]
-        });
-        let out = jq(&filter, &val).unwrap();
-        assert!(out.contains(r#""missing" (unresolved)"#));
-    }
-
-    // --- FileObject text rendering through format_value_as_text ---
-
-    #[test]
-    fn file_object_text_rendering_minimal() {
-        let val = json!({"file": "notes/foo.md", "modified": "2024-01-15"});
-        let out = fmt(&val);
-        assert!(out.contains("notes/foo.md"));
-        assert!(out.contains("2024-01-15"));
-        // Should not look like generic fallback
-        assert!(!out.contains("file: notes/foo.md"));
-    }
-
-    #[test]
-    fn file_object_text_rendering_full() {
-        let val = json!({
-            "file": "notes/project.md",
-            "modified": "2024-03-01",
-            "tags": ["rust", "work"],
-            "properties": {"status": "active"},
-            "tasks": [
-                {"done": false, "line": 10, "section": "Todo", "status": " ", "text": "Fix bug"},
-                {"done": true, "line": 20, "section": "Done", "status": "x", "text": "Write docs"}
-            ]
-        });
-        let out = fmt(&val);
-        assert!(out.contains("notes/project.md"));
-        assert!(out.contains("properties:"));
-        assert!(out.contains("status: active"));
-        assert!(out.contains("tags: [rust, work]"));
-        assert!(out.contains("tasks:"));
-        assert!(out.contains("[ ] Fix bug"));
-        assert!(out.contains("[x] Write docs"));
-    }
-
-    // --- Array of FileObjects with blank-line separator ---
-
-    #[test]
-    fn array_of_file_objects_uses_blank_line_separator() {
-        let val = json!([
-            {"file": "a.md", "modified": "2024-01-01"},
-            {"file": "b.md", "modified": "2024-01-02"}
-        ]);
-        let out = fmt(&val);
-        assert!(out.contains("a.md"));
-        assert!(out.contains("b.md"));
-        // Should have a blank line between entries
-        assert!(
-            out.contains("\n\n"),
-            "expected blank-line separator between file objects"
-        );
-    }
-
-    #[test]
-    fn array_of_non_file_objects_uses_single_newline() {
-        let val = json!([
-            {"count": 1, "name": "status", "type": "text"},
-            {"count": 3, "name": "title", "type": "text"}
-        ]);
-        let out = fmt(&val);
-        assert!(out.contains("status"));
-        assert!(out.contains("title"));
-        // Should NOT have a blank line separator
-        assert!(
-            !out.contains("\n\n"),
-            "non-file-objects should use single newline"
-        );
-    }
-
-    // --- format_scalar nested object delegation ---
-
-    #[test]
-    fn format_scalar_delegates_nested_objects() {
-        // A nested object with a known shape should get its filter applied,
-        // not the k=v flat format.
-        let inner = json!({"count": 2, "name": "status", "type": "text"});
-        let out = scalar(&inner);
-        // Should NOT look like the old "count=2, name=status, type=text" format.
-        assert!(
-            !out.contains("count=2"),
-            "should delegate to format_value_as_text"
-        );
-        // Should look like the PropertySummaryEntry filter output.
-        assert!(out.contains("status"));
-        assert!(out.contains("2 files"));
-    }
-
-    // --- format_value_as_text integration ---
-
-    #[test]
-    fn format_value_as_text_uses_filter_for_known_shape() {
-        // PropertySummaryEntry has a known shape: {count, name, type}
-        let val = json!({"count": 3, "name": "status", "type": "text"});
-        let out = fmt(&val);
-        assert!(out.contains("status"));
-        assert!(out.contains("3 files"));
-        // Should NOT look like "count: 3" (that's the generic fallback)
-        assert!(!out.contains("count: 3"));
-    }
-
-    #[test]
-    fn format_value_as_text_falls_back_for_unknown_shape() {
-        let val = json!({"foo": "bar", "baz": 42});
-        let out = fmt(&val);
-        // Generic fallback: key: value
-        assert!(out.contains("foo: bar") || out.contains("baz: 42"));
-    }
-
-    #[test]
-    fn mv_result_filter_applied() {
-        let val = json!({
-            "dry_run": false,
-            "from": "sub/b.md",
-            "to": "archive/b.md",
-            "total_files_updated": 1,
-            "total_links_updated": 1,
-            "updated_files": [
-                {
-                    "file": "a.md",
-                    "replacements": [
-                        {"old_text": "[[sub/b]]", "new_text": "[[archive/b]]", "line": 1}
-                    ]
-                }
-            ]
-        });
-        // Verify key signature matches expected
-        let sig = {
-            let map = val.as_object().unwrap();
-            let mut keys: Vec<&str> = map.keys().map(String::as_str).collect();
-            keys.sort_unstable();
-            keys.join(",")
-        };
-        assert_eq!(
-            sig,
-            "dry_run,from,to,total_files_updated,total_links_updated,updated_files"
-        );
-        // Verify the jq filter itself works
-        let filter_result = apply_jq_filter_result(MV_RESULT_FILTER, &val);
-        assert!(filter_result.is_ok(), "filter error: {filter_result:?}");
-        let out = filter_result.unwrap();
-        assert!(out.contains("Moved sub/b.md"), "out: {out}");
-        assert!(out.contains("archive/b.md"), "out: {out}");
-        assert!(out.contains("[[sub/b]]"), "out: {out}");
-        assert!(out.contains("[[archive/b]]"), "out: {out}");
-        // Verify lookup_filter finds the filter for this shape
-        let found_filter =
-            lookup_filter("dry_run,from,to,total_files_updated,total_links_updated,updated_files");
-        assert!(
-            found_filter.is_some(),
-            "lookup_filter returned None for MvResult shape"
-        );
-        // format_value_as_text should pick up the filter
-        let formatted = fmt(&val);
-        assert!(
-            formatted.contains("Moved sub/b.md"),
-            "formatted: {formatted}"
-        );
-    }
-
-    #[test]
-    fn format_value_as_text_array_of_typed_objects() {
-        let val = json!([
-            {"path": "a.md", "tags": ["rust"]},
-            {"path": "b.md", "tags": ["cli"]}
-        ]);
-        let out = fmt(&val);
-        assert!(out.contains("a.md"));
-        assert!(out.contains("b.md"));
-        assert!(out.contains("rust"));
-        assert!(out.contains("cli"));
-    }
-
-    // --- sanitize_control_chars ---
-
-    #[test]
-    fn sanitize_control_chars_strips_escape_sequences() {
-        let input = "Hello\x1b[31mRED\x1b[0m World";
-        let output = sanitize_control_chars(input);
-        assert!(
-            !output.contains('\x1b'),
-            "escape sequences should be stripped"
-        );
-        assert!(output.contains("Hello"));
-        assert!(output.contains("RED"));
-        assert!(output.contains("World"));
-    }
-
-    #[test]
-    fn sanitize_control_chars_preserves_newline_and_tab() {
-        let input = "line1\nline2\ttabbed";
-        let output = sanitize_control_chars(input);
-        assert_eq!(output, input);
-    }
-
-    #[test]
-    fn text_output_sanitizes_escape_sequences() {
-        let value = serde_json::json!({
-            "results": {
-                "title": "Hello\x1b[31mRED\x1b[0m World",
-                "file": "test\x1b[2J.md"
-            }
-        });
-        let output = format_success(Format::Text, &value);
-        assert!(
-            !output.contains('\x1b'),
-            "escape sequences should be stripped"
-        );
-        assert!(output.contains("Hello") && output.contains("World"));
-    }
-
-    // --- UX-6: lint-rules mutation text renderer ---
-
-    /// `lint-rules set` with enabled change renders before→after and config path.
-    #[test]
-    fn lint_rules_set_enabled_change_renders_diff() {
-        let value = serde_json::json!({
-            "action": "set",
-            "rule_id": "MD013",
-            "dry_run": false,
-            "before": {"enabled": true, "severity": "warn"},
-            "after": {"enabled": false, "severity": "warn"},
-            "config_path": ".hyalo.toml"
-        });
-        let out = format_success(Format::Text, &value);
-        assert!(out.contains("MD013:"), "should include rule id");
-        assert!(
-            out.contains("on") && out.contains("off"),
-            "should show enabled change"
-        );
-        assert!(out.contains("wrote .hyalo.toml"), "should mention write");
-    }
-
-    /// `lint-rules set` with severity change only.
-    #[test]
-    fn lint_rules_set_severity_change_renders_diff() {
-        let value = serde_json::json!({
-            "action": "set",
-            "rule_id": "HYALO001",
-            "dry_run": false,
-            "before": {"enabled": true, "severity": "warn"},
-            "after": {"enabled": true, "severity": "error"},
-            "config_path": ".hyalo.toml"
-        });
-        let out = format_success(Format::Text, &value);
-        assert!(out.contains("HYALO001:"));
-        assert!(out.contains("warn") && out.contains("error"));
-        assert!(out.contains("wrote .hyalo.toml"));
-    }
-
-    /// `lint-rules set` with dry-run says "dry-run, would write".
-    #[test]
-    fn lint_rules_set_dry_run_says_would_write() {
-        let value = serde_json::json!({
-            "action": "set",
-            "rule_id": "MD013",
-            "dry_run": true,
-            "before": {"enabled": true, "severity": "warn"},
-            "after": {"enabled": false, "severity": "warn"},
-            "config_path": ".hyalo.toml"
-        });
-        let out = format_success(Format::Text, &value);
-        assert!(out.contains("dry-run"), "should mention dry-run");
-        assert!(!out.contains("wrote "), "should not say wrote");
-    }
-
-    /// `lint-rules remove` with a removed override shows reverted state.
-    #[test]
-    fn lint_rules_remove_shows_reverted_state() {
-        let value = serde_json::json!({
-            "action": "remove",
-            "rule_id": "MD013",
-            "dry_run": false,
-            "removed": true,
-            "before": {"enabled": false, "severity": "warn"},
-            "after": {"enabled": true, "severity": "warn"},
-            "config_path": ".hyalo.toml"
-        });
-        let out = format_success(Format::Text, &value);
-        assert!(out.contains("MD013:"));
-        assert!(out.contains("removed override"));
-        assert!(out.contains("wrote .hyalo.toml"));
-    }
-
-    /// `lint-rules remove` when nothing to remove shows no-op message.
-    #[test]
-    fn lint_rules_remove_noop_shows_reason() {
-        let value = serde_json::json!({
-            "action": "remove",
-            "rule_id": "MD013",
-            "dry_run": false,
-            "removed": false,
-            "reason": "no override found",
-            "before": {"enabled": true, "severity": "warn"},
-            "after": {"enabled": true, "severity": "warn"},
-            "config_path": ".hyalo.toml"
-        });
-        let out = format_success(Format::Text, &value);
-        assert!(out.contains("MD013:"));
-        assert!(out.contains("no override to remove") || out.contains("no override found"));
-    }
-
-    // --- build_envelope_value: NEW-5 dir dedup ---
-
-    #[test]
-    fn envelope_hoists_dir_and_removes_from_results() {
-        let results = json!({
-            "dir": "hyalo-knowledgebase",
-            "files": {"total": 10},
-            "other": "value"
-        });
-        let envelope = build_envelope_value(&results, None, &[]);
-        // dir is hoisted to top level.
-        assert_eq!(
-            envelope["dir"].as_str().unwrap(),
-            "hyalo-knowledgebase",
-            "dir must be at envelope root"
-        );
-        // dir is removed from results.
-        assert!(
-            envelope["results"]
-                .get("dir")
-                .is_none_or(serde_json::Value::is_null),
-            "dir must be removed from results; envelope: {envelope}"
-        );
-        // Other results fields are preserved.
-        assert_eq!(envelope["results"]["files"]["total"].as_u64().unwrap(), 10);
-        assert_eq!(envelope["results"]["other"].as_str().unwrap(), "value");
-    }
-
-    #[test]
-    fn envelope_without_dir_in_results_has_no_top_dir() {
-        let results = json!({"files": {"total": 5}});
-        let envelope = build_envelope_value(&results, None, &[]);
-        assert!(
-            envelope.get("dir").is_none() || envelope["dir"].is_null(),
-            "no dir should be at envelope root when results has none"
-        );
-    }
-}
+mod tests;
