+++ crates/hyalo-cli/src/output/text_lint.rs
@@ -1,0 +1,793 @@
+//! Hand-written text renderings for `lint`, `lint --fix` and `lint-rules`.
+//!
+//! Split out of the single 3,744-line `output.rs` in iteration 247
+//! (deep-review hotspot). A file split only: every item keeps the visibility it
+//! had in the one module, so `output::...` paths and behaviour are unchanged.
+
+/// Format `ExtLintOutput` JSON as human-readable text.
+///
+/// Reproduces the format previously generated by `commands::lint::format_text_output`.
+pub(super) fn format_lint_output_text(map: &serde_json::Map<String, serde_json::Value>) -> String {
+    use std::fmt::Write as _;
+
+    let mut s = String::new();
+    let dry_run = map
+        .get("dry_run")
+        .and_then(serde_json::Value::as_bool)
+        .unwrap_or(false);
+
+    // Fix actions (shown first).
+    if let Some(fixes_arr) = map.get("fixes").and_then(|f| f.as_array()) {
+        let verb = if dry_run { "Would fix" } else { "Fixed" };
+        for file_fix in fixes_arr {
+            let file = file_fix
+                .get("file")
+                .and_then(serde_json::Value::as_str)
+                .unwrap_or("?");
+            let actions = file_fix.get("actions").and_then(|a| a.as_array());
+            let Some(actions) = actions else { continue };
+            if actions.is_empty() {
+                continue;
+            }
+            let _ = writeln!(s, "{verb} {file}:");
+            for a in actions {
+                let kind = a
+                    .get("kind")
+                    .and_then(serde_json::Value::as_str)
+                    .unwrap_or("");
+                let property = a
+                    .get("property")
+                    .and_then(serde_json::Value::as_str)
+                    .unwrap_or("?");
+                let new = a
+                    .get("new")
+                    .and_then(serde_json::Value::as_str)
+                    .unwrap_or("");
+                let old = a.get("old").and_then(serde_json::Value::as_str);
+                match (kind, old) {
+                    ("insert-default", _) => {
+                        let _ = writeln!(s, "  insert  {property} = {new:?}");
+                    }
+                    ("infer-type", _) => {
+                        let _ = writeln!(s, "  infer   type = {new:?}");
+                    }
+                    ("fix-enum-typo", Some(old_v)) => {
+                        let _ = writeln!(s, "  enum    {property}: {old_v:?} -> {new:?}");
+                    }
+                    ("normalize-date", Some(old_v)) => {
+                        let _ = writeln!(s, "  date    {property}: {old_v:?} -> {new:?}");
+                    }
+                    _ => {
+                        let _ = writeln!(s, "  {kind}  {property} = {new:?}");
+                    }
+                }
+            }
+        }
+    }
+
+    // File violations — handle both the new `rule_groups` shape and the legacy `violations` shape.
+    let files = map.get("files").and_then(|f| f.as_array());
+    if let Some(files) = files {
+        for file_entry in files {
+            let file = file_entry
+                .get("file")
+                .and_then(serde_json::Value::as_str)
+                .unwrap_or("?");
+
+            // New shape: violations are grouped by rule under `rule_groups`.
+            if let Some(rule_groups) = file_entry.get("rule_groups").and_then(|rg| rg.as_array()) {
+                if rule_groups.is_empty() {
+                    continue;
+                }
+                let _ = writeln!(s, "{file}:");
+                for group in rule_groups {
+                    let rule = group
+                        .get("rule")
+                        .and_then(serde_json::Value::as_str)
+                        .unwrap_or("?");
+                    // Group severity is the fallback when a violation does not
+                    // carry its own (older payloads); a folded group such as
+                    // `SCHEMA` mixes severities, so each line is labelled from
+                    // its own `severity` field to agree with the counts (BUG-17).
+                    let group_severity = group
+                        .get("severity")
+                        .and_then(serde_json::Value::as_str)
+                        .unwrap_or("warn");
+                    let violations = group.get("violations").and_then(|v| v.as_array());
+                    let count = group
+                        .get("count")
+                        .and_then(serde_json::Value::as_u64)
+                        .unwrap_or(0);
+                    let shown = group
+                        .get("shown")
+                        .and_then(serde_json::Value::as_u64)
+                        .unwrap_or(count);
+                    if let Some(violations) = violations {
+                        for v in violations {
+                            let line = v
+                                .get("line")
+                                .and_then(serde_json::Value::as_u64)
+                                .unwrap_or(0);
+                            let message = v
+                                .get("message")
+                                .and_then(serde_json::Value::as_str)
+                                .unwrap_or("");
+                            let severity = v
+                                .get("severity")
+                                .and_then(serde_json::Value::as_str)
+                                .unwrap_or(group_severity);
+                            let pad = if severity == "error" {
+                                "error"
+                            } else {
+                                "warn "
+                            };
+                            if line > 0 {
+                                let _ = writeln!(s, "  {pad}  {rule}  line {line}  {message}");
+                            } else {
+                                let _ = writeln!(s, "  {pad}  {rule}  {message}");
+                            }
+                        }
+                        if count > shown {
+                            let _ = writeln!(s, "  … ({} more {})", count - shown, rule);
+                        }
+                    }
+                }
+            } else if let Some(violations) = file_entry.get("violations").and_then(|v| v.as_array())
+            {
+                // Legacy shape: flat violations array.
+                if violations.is_empty() {
+                    continue;
+                }
+                let _ = writeln!(s, "{file}:");
+                for v in violations {
+                    let severity = v
+                        .get("severity")
+                        .and_then(serde_json::Value::as_str)
+                        .unwrap_or("warn");
+                    let message = v
+                        .get("message")
+                        .and_then(serde_json::Value::as_str)
+                        .unwrap_or("");
+                    let pad = if severity == "error" {
+                        "error"
+                    } else {
+                        "warn "
+                    };
+                    let _ = writeln!(s, "  {pad}  {message}");
+                }
+            }
+        }
+    }
+
+    let error_count: u64 = map
+        .get("errors")
+        .and_then(serde_json::Value::as_u64)
+        .unwrap_or(0);
+    let warn_count: u64 = map
+        .get("warnings")
+        .and_then(serde_json::Value::as_u64)
+        .unwrap_or(0);
+    // Support both old `files_with_issues` and new `files_with_violations`.
+    let files_with_issues: u64 = map
+        .get("files_with_violations")
+        .or_else(|| map.get("files_with_issues"))
+        .and_then(serde_json::Value::as_u64)
+        .unwrap_or(0);
+    // Support both old `limited` and new `files_truncated`.
+    let limited = map
+        .get("files_truncated")
+        .or_else(|| map.get("limited"))
+        .and_then(serde_json::Value::as_bool)
+        .unwrap_or(false);
+    let shown_files = map
+        .get("files")
+        .and_then(|f| f.as_array())
+        .map_or(0, |arr| {
+            arr.iter()
+                .filter(|e| {
+                    // New shape: has rule_groups
+                    e.get("rule_groups")
+                        .and_then(|rg| rg.as_array())
+                        .is_some_and(|rg| !rg.is_empty())
+                        // Legacy shape: has violations
+                        || e.get("violations")
+                            .and_then(|v| v.as_array())
+                            .is_some_and(|v| !v.is_empty())
+                })
+                .count()
+        });
+
+    if limited {
+        let _ = writeln!(
+            s,
+            "… (showing {shown_files} of {files_with_issues} files with issues)"
+        );
+    }
+
+    // Summary line.
+    let files_checked: u64 = map
+        .get("files_checked")
+        .and_then(serde_json::Value::as_u64)
+        .unwrap_or(0);
+    let files_label = if files_checked == 1 { "file" } else { "files" };
+    if error_count == 0 && warn_count == 0 {
+        let _ = write!(s, "{files_checked} {files_label} checked, no issues");
+    } else {
+        let errors_label = if error_count == 1 { "error" } else { "errors" };
+        let warns_label = if warn_count == 1 {
+            "warning"
+        } else {
+            "warnings"
+        };
+        let _ = write!(
+            s,
+            "{files_checked} {files_label} checked, {files_with_issues} with issues ({error_count} {errors_label}, {warn_count} {warns_label})",
+        );
+    }
+    // UX-1 (dogfood pre3): a bare sweep silently dropped every `[lint]
+    // ignore`-matched file with no visible trace — "68 files checked, no
+    // issues" on a 386-file vault read as a clean bill of health, not "68 of
+    // 386 were actually looked at." The named-file / glob-all-ignored cases
+    // already get their own stderr notice; this is the everything-else case.
+    let files_ignored: u64 = map
+        .get("files_ignored")
+        .and_then(serde_json::Value::as_u64)
+        .unwrap_or(0);
+    if files_ignored > 0 {
+        let _ = write!(s, " ({files_ignored} ignored by [lint] ignore)");
+    }
+
+    let fix_count: usize = map
+        .get("fixes")
+        .and_then(|f| f.as_array())
+        .map_or(0, |arr| {
+            arr.iter()
+                .filter_map(|f| f.get("actions").and_then(|a| a.as_array()).map(Vec::len))
+                .sum()
+        });
+    if fix_count > 0 {
+        let fixed_label = if dry_run { "would fix" } else { "fixed" };
+        let _ = write!(s, " — {fixed_label} {fix_count}");
+    }
+
+    s
+}
+
+/// Format `ExtLintFixOutput` (fix-mode lint) as human-readable text.
+///
+/// Renders three prefixes: `fixed`, `remain`, `conflict`.
+pub(super) fn format_lint_fix_output_text(
+    map: &serde_json::Map<String, serde_json::Value>,
+) -> String {
+    use std::fmt::Write as _;
+
+    let mut s = String::new();
+    let dry_run = map
+        .get("dry_run")
+        .and_then(serde_json::Value::as_bool)
+        .unwrap_or(false);
+
+    let files = map.get("files").and_then(|f| f.as_array());
+    if let Some(files) = files {
+        for file_entry in files {
+            let file = file_entry
+                .get("file")
+                .and_then(serde_json::Value::as_str)
+                .unwrap_or("?");
+
+            let has_content = file_entry
+                .get("fixed_groups")
+                .and_then(|g| g.as_array())
+                .is_some_and(|a| !a.is_empty())
+                || file_entry
+                    .get("remaining_groups")
+                    .and_then(|g| g.as_array())
+                    .is_some_and(|a| !a.is_empty())
+                || file_entry
+                    .get("conflicts")
+                    .and_then(|c| c.as_array())
+                    .is_some_and(|a| !a.is_empty());
+            if !has_content {
+                continue;
+            }
+
+            let _ = writeln!(s, "{file}:");
+            let prefix = if dry_run { "would fix" } else { "fixed  " };
+
+            // Fixed groups.
+            if let Some(fixed_groups) = file_entry.get("fixed_groups").and_then(|g| g.as_array()) {
+                for group in fixed_groups {
+                    let rule = group
+                        .get("rule")
+                        .and_then(serde_json::Value::as_str)
+                        .unwrap_or("?");
+                    let violations = group.get("violations").and_then(|v| v.as_array());
+                    let count = group
+                        .get("count")
+                        .and_then(serde_json::Value::as_u64)
+                        .unwrap_or(0);
+                    if let Some(violations) = violations.filter(|v| !v.is_empty()) {
+                        for v in violations {
+                            let line = v
+                                .get("line")
+                                .and_then(serde_json::Value::as_u64)
+                                .unwrap_or(0);
+                            let message = v
+                                .get("message")
+                                .and_then(serde_json::Value::as_str)
+                                .unwrap_or("");
+                            if line > 0 {
+                                let _ = writeln!(s, "  {prefix}  {rule}  line {line}  {message}");
+                            } else {
+                                let _ = writeln!(s, "  {prefix}  {rule}  {message}");
+                            }
+                        }
+                    } else if count > 0 {
+                        let _ = writeln!(s, "  {prefix}  {rule}  ({count} violations)");
+                    }
+                }
+            }
+
+            // Remaining groups.
+            if let Some(remaining_groups) = file_entry
+                .get("remaining_groups")
+                .and_then(|g| g.as_array())
+            {
+                for group in remaining_groups {
+                    let rule = group
+                        .get("rule")
+                        .and_then(serde_json::Value::as_str)
+                        .unwrap_or("?");
+                    let severity = group
+                        .get("severity")
+                        .and_then(serde_json::Value::as_str)
+                        .unwrap_or("warn");
+                    let sev_pad = if severity == "error" {
+                        "error"
+                    } else {
+                        "warn "
+                    };
+                    let violations = group.get("violations").and_then(|v| v.as_array());
+                    let count = group
+                        .get("count")
+                        .and_then(serde_json::Value::as_u64)
+                        .unwrap_or(0);
+                    let shown = group
+                        .get("shown")
+                        .and_then(serde_json::Value::as_u64)
+                        .unwrap_or(count);
+                    if let Some(violations) = violations {
+                        for v in violations {
+                            let line = v
+                                .get("line")
+                                .and_then(serde_json::Value::as_u64)
+                                .unwrap_or(0);
+                            let message = v
+                                .get("message")
+                                .and_then(serde_json::Value::as_str)
+                                .unwrap_or("");
+                            if line > 0 {
+                                let _ = writeln!(
+                                    s,
+                                    "  remain  {sev_pad}  {rule}  line {line}  {message}"
+                                );
+                            } else {
+                                let _ = writeln!(s, "  remain  {sev_pad}  {rule}  {message}");
+                            }
+                        }
+                        if count > shown {
+                            let _ = writeln!(s, "  … ({} more {})", count - shown, rule);
+                        }
+                    }
+                }
+            }
+
+            // Conflicts.
+            //
+            // A rule with several violations in one file can have some fixes
+            // applied and one lose a range overlap, which printed the rule as
+            // both `fixed` and `conflict` two lines apart — read as a
+            // contradiction rather than as two different violations (iter-213,
+            // dogfood UX-5). The `fixed` line is the one that describes what
+            // changed on disk, so the conflict line is dropped when the same
+            // rule already appears there. The JSON keeps both: `conflicts` is
+            // how a consumer learns a fix was skipped, and dropping entries
+            // from it would lose that.
+            let fixed_rules: std::collections::HashSet<&str> = file_entry
+                .get("fixed_groups")
+                .and_then(|g| g.as_array())
+                .map(|groups| {
+                    groups
+                        .iter()
+                        .filter_map(|g| g.get("rule").and_then(serde_json::Value::as_str))
+                        .collect()
+                })
+                .unwrap_or_default();
+            if let Some(conflicts) = file_entry.get("conflicts").and_then(|c| c.as_array()) {
+                for conflict in conflicts {
+                    let rule = conflict
+                        .get("rule")
+                        .and_then(serde_json::Value::as_str)
+                        .unwrap_or("?");
+                    if fixed_rules.contains(rule) {
+                        continue;
+                    }
+                    let reason = conflict
+                        .get("reason")
+                        .and_then(serde_json::Value::as_str)
+                        .unwrap_or("");
+                    let _ = writeln!(s, "  conflict  {rule}  {reason}");
+                }
+            }
+        }
+    }
+
+    // `files_truncated` gates a "(showing N of M files with issues)" line,
+    // mirroring `format_lint_output_text`'s. Without it, the summary line
+    // below — whose fixed/remaining/conflicts counts describe the WHOLE run
+    // since iter-218 (NEW-6) — read as if the file listing above it were
+    // complete too, when `--limit` had silently capped it (review finding
+    // #2).
+    let files_truncated = map
+        .get("files_truncated")
+        .and_then(serde_json::Value::as_bool)
+        .unwrap_or(false);
+    if files_truncated {
+        let files_with_violations = map
+            .get("files_with_violations")
+            .and_then(serde_json::Value::as_u64)
+            .unwrap_or(0);
+        let shown_files = files.map_or(0, |arr| {
+            arr.iter()
+                .filter(|f| {
+                    f.get("fixed_groups")
+                        .and_then(|g| g.as_array())
+                        .is_some_and(|a| !a.is_empty())
+                        || f.get("remaining_groups")
+                            .and_then(|g| g.as_array())
+                            .is_some_and(|a| !a.is_empty())
+                        || f.get("conflicts")
+                            .and_then(|c| c.as_array())
+                            .is_some_and(|a| !a.is_empty())
+                })
+                .count()
+        });
+        let _ = writeln!(
+            s,
+            "… (showing {shown_files} of {files_with_violations} files with issues)"
+        );
+    }
+
+    // Summary line.
+    let files_checked = map
+        .get("files_checked")
+        .and_then(serde_json::Value::as_u64)
+        .unwrap_or(0);
+    let total_fixed = map
+        .get("total_fixed")
+        .and_then(serde_json::Value::as_u64)
+        .unwrap_or(0);
+    let total_remaining = map
+        .get("total_remaining")
+        .and_then(serde_json::Value::as_u64)
+        .unwrap_or(0);
+    let total_conflicts = map
+        .get("total_conflicts")
+        .and_then(serde_json::Value::as_u64)
+        .unwrap_or(0);
+    let files_label = if files_checked == 1 { "file" } else { "files" };
+    let fixed_phrase = if dry_run {
+        format!("would fix {total_fixed}")
+    } else {
+        format!("fixed {total_fixed}")
+    };
+    let _ = write!(
+        s,
+        "{files_checked} {files_label} checked: {fixed_phrase} · remaining {total_remaining} · conflicts {total_conflicts}.",
+    );
+
+    s
+}
+
+/// Format a `lint-rules list` result as a column table.
+///
+/// Detects arrays of objects with `id`, `default_severity`, `effective_enabled`,
+/// `autofixable`, `source` fields.
+pub(super) fn format_lint_rules_list_text(arr: &[serde_json::Value]) -> String {
+    use std::fmt::Write as _;
+
+    const DEFAULT_TERM_WIDTH: usize = 120;
+
+    if arr.is_empty() {
+        return String::new();
+    }
+
+    // Verify it looks like lint-rule entries.
+    let first_is_rule = arr.first().and_then(|v| v.as_object()).is_some_and(|m| {
+        m.contains_key("id")
+            && m.contains_key("effective_enabled")
+            && m.contains_key("autofixable")
+            && m.contains_key("source")
+    });
+    if !first_is_rule {
+        return String::new();
+    }
+
+    let term_width = DEFAULT_TERM_WIDTH;
+
+    // Collect rows.
+    let mut rows: Vec<[String; 6]> = Vec::with_capacity(arr.len());
+    let mut total_enabled = 0usize;
+    let mut total_autofixable = 0usize;
+
+    for entry in arr {
+        let Some(obj) = entry.as_object() else {
+            continue;
+        };
+        let id = obj
+            .get("id")
+            .and_then(serde_json::Value::as_str)
+            .unwrap_or("")
+            .to_owned();
+        let name = obj
+            .get("name")
+            .and_then(serde_json::Value::as_str)
+            .unwrap_or("")
+            .to_owned();
+        let severity = obj
+            .get("effective_severity")
+            .or_else(|| obj.get("default_severity"))
+            .and_then(serde_json::Value::as_str)
+            .unwrap_or("warn")
+            .to_owned();
+        let enabled = obj
+            .get("effective_enabled")
+            .and_then(serde_json::Value::as_bool)
+            .unwrap_or(false);
+        let autofix = obj
+            .get("autofixable")
+            .and_then(serde_json::Value::as_bool)
+            .unwrap_or(false);
+        let description = obj
+            .get("description")
+            .and_then(serde_json::Value::as_str)
+            .unwrap_or("")
+            .to_owned();
+
+        if enabled {
+            total_enabled += 1;
+        }
+        if autofix {
+            total_autofixable += 1;
+        }
+
+        rows.push([
+            id,
+            name,
+            severity,
+            if enabled {
+                "on".to_owned()
+            } else {
+                "off".to_owned()
+            },
+            if autofix {
+                "yes".to_owned()
+            } else {
+                "no".to_owned()
+            },
+            description,
+        ]);
+    }
+
+    // Column headers.
+    let headers = [
+        "ID",
+        "NAME",
+        "SEVERITY",
+        "ENABLED",
+        "AUTOFIX",
+        "DESCRIPTION",
+    ];
+
+    // Compute max widths for columns 0-4 (description is last, capped).
+    let mut col_widths = [0usize; 6];
+    for (i, h) in headers.iter().enumerate() {
+        col_widths[i] = h.len();
+    }
+    for row in &rows {
+        for i in 0..5 {
+            col_widths[i] = col_widths[i].max(row[i].len());
+        }
+    }
+
+    // Compute remaining width for description column.
+    // Fixed columns + separators (2 spaces between each).
+    let fixed_width: usize = col_widths[..5].iter().sum::<usize>() + 5 * 2; // 5 gaps of 2 spaces
+    let desc_available = term_width.saturating_sub(fixed_width).max(20);
+    let desc_width = desc_available.min(80);
+    col_widths[5] = desc_width;
+
+    let mut s = String::new();
+
+    // Header row.
+    for (i, h) in headers.iter().enumerate() {
+        if i > 0 {
+            s.push_str("  ");
+        }
+        if i == 5 {
+            s.push_str(h);
+        } else {
+            let _ = write!(s, "{h:<width$}", width = col_widths[i]);
+        }
+    }
+    s.push('\n');
+
+    // Data rows.
+    for row in &rows {
+        for (i, cell) in row.iter().enumerate() {
+            if i > 0 {
+                s.push_str("  ");
+            }
+            if i == 5 {
+                // Truncate description with ellipsis if needed.
+                if cell.chars().count() > desc_width {
+                    let truncated: String =
+                        cell.chars().take(desc_width.saturating_sub(1)).collect();
+                    let _ = write!(s, "{truncated}…");
+                } else {
+                    s.push_str(cell);
+                }
+            } else {
+                let _ = write!(s, "{cell:<width$}", width = col_widths[i]);
+            }
+        }
+        s.push('\n');
+    }
+
+    // Summary line.
+    let total = arr.len();
+    let total_disabled = total - total_enabled;
+    let _ = write!(
+        s,
+        "\n{total} rules, {total_enabled} enabled, {total_disabled} disabled. {total_autofixable} autofixable."
+    );
+
+    s
+}
+
+/// Map `Some(true)` → `"on"`, `Some(false)` → `"off"`, `None` → `"?"`.
+pub(super) fn bool_label(b: Option<bool>) -> &'static str {
+    match b {
+        Some(true) => "on",
+        Some(false) => "off",
+        None => "?",
+    }
+}
+
+/// Format a `lint-rules set` or `lint-rules remove` result as human-readable text.
+///
+/// Expected JSON shapes:
+/// - set:    `{action: "set",    rule_id, dry_run, before: {enabled, severity}, after: {enabled, severity}, config_path}`
+/// - remove: `{action: "remove", rule_id, dry_run, removed, before: {enabled, severity}, after: {enabled, severity}, config_path}`
+///
+/// Rendered example:
+/// ```text
+/// HYALO002: enabled on → off (warn → warn)
+///   wrote .hyalo.toml
+/// ```
+pub(super) fn format_lint_rules_mutation_text(
+    map: &serde_json::Map<String, serde_json::Value>,
+) -> String {
+    use std::fmt::Write as _;
+
+    let action = map
+        .get("action")
+        .and_then(serde_json::Value::as_str)
+        .unwrap_or("set");
+    let rule_id = map
+        .get("rule_id")
+        .and_then(serde_json::Value::as_str)
+        .unwrap_or("?");
+    let dry_run = map
+        .get("dry_run")
+        .and_then(serde_json::Value::as_bool)
+        .unwrap_or(false);
+    let config_path = map
+        .get("config_path")
+        .and_then(serde_json::Value::as_str)
+        .unwrap_or(".hyalo.toml");
+
+    let before_enabled = map
+        .get("before")
+        .and_then(|b| b.get("enabled"))
+        .and_then(serde_json::Value::as_bool);
+    let before_severity = map
+        .get("before")
+        .and_then(|b| b.get("severity"))
+        .and_then(serde_json::Value::as_str)
+        .unwrap_or("warn");
+    let after_enabled = map
+        .get("after")
+        .and_then(|b| b.get("enabled"))
+        .and_then(serde_json::Value::as_bool);
+    let after_severity = map
+        .get("after")
+        .and_then(|b| b.get("severity"))
+        .and_then(serde_json::Value::as_str)
+        .unwrap_or("warn");
+
+    let mut s = String::new();
+
+    if action == "remove" {
+        let removed = map
+            .get("removed")
+            .and_then(serde_json::Value::as_bool)
+            .unwrap_or(false);
+        if !removed {
+            let reason = map
+                .get("reason")
+                .and_then(serde_json::Value::as_str)
+                .unwrap_or("no override found");
+            let _ = write!(s, "{rule_id}: no override to remove ({reason})");
+            return s;
+        }
+        // Show what changed when removing (reverts to defaults).
+        let en_changed = before_enabled != after_enabled;
+        let sev_changed = before_severity != after_severity;
+        let _ = write!(s, "{rule_id}: removed override");
+        if en_changed || sev_changed {
+            let _ = write!(s, " ({}", bool_label(before_enabled));
+            if sev_changed {
+                let _ = write!(s, ", {before_severity}");
+            }
+            let _ = write!(s, " → {}", bool_label(after_enabled));
+            if sev_changed {
+                let _ = write!(s, ", {after_severity}");
+            }
+            s.push(')');
+        }
+    } else {
+        // set action: show enabled and severity changes side-by-side.
+        let en_changed = before_enabled != after_enabled;
+        let sev_changed = before_severity != after_severity;
+
+        let _ = write!(s, "{rule_id}:");
+        if en_changed {
+            let _ = write!(
+                s,
+                " enabled {} → {}",
+                bool_label(before_enabled),
+                bool_label(after_enabled)
+            );
+        }
+        if sev_changed {
+            if en_changed {
+                s.push(',');
+            }
+            let _ = write!(s, " {before_severity} → {after_severity}");
+        }
+        if !en_changed && !sev_changed {
+            s.push_str(" (no change)");
+        }
+    }
+
+    // Indicate write vs dry-run vs no-op.
+    let wrote = map
+        .get("wrote")
+        .and_then(serde_json::Value::as_bool)
+        .unwrap_or(true);
+    s.push('\n');
+    if dry_run {
+        if wrote {
+            let _ = write!(s, "  dry-run, would write {config_path}");
+        } else {
+            let _ = write!(s, "  dry-run, no changes to {config_path}");
+        }
+    } else if wrote {
+        let _ = write!(s, "  wrote {config_path}");
+    } else {
+        let _ = write!(s, "  {config_path} unchanged");
+    }
+
+    s
+}
